@@ -1,21 +1,17 @@
+requires-python = "==3.11.*" # Required by bpy 3.6.*
+version = "1.0.0" # Dummy version. See __init__.bl_info["version"] instead.
+authors = [
+    {name = "Isamu Mogi"},
+    {name = "iCyP"},
+]
+license = {text = "SPDX-License-Identifier: MIT OR GPL-3.0-or-later"}
+dependencies = [
+    "bpy>=4.5.3",
+]
 [project]
 name = "io_scene_vrm"
 version = "3.13.1"
-authors = [
-  {name = "Isamu Mogi"},
-  {name = "iCyP"},
-]
-license = {text = "SPDX-License-Identifier: MIT OR GPL-3.0-or-later"}
-
-requires-python = "==3.10.*" # I actually want to use 3.9, but bpy 3.6.* doesn't allow it.
-
-dependencies = [
-  # https://github.com/saturday06/VRM-Addon-for-Blender/blob/v3.10.1/uv.lock#L12-L27
-  "bpy ==3.6.*; platform_machine == 'arm64' and sys_platform == 'darwin'",
-  "bpy ==3.6.*; platform_machine == 'x86_64' and sys_platform == 'darwin'",
-  "bpy ==3.6.*; platform_machine == 'x86_64' and sys_platform == 'linux'",
-  "bpy ==3.6.*; platform_machine == 'AMD64' and sys_platform == 'win32'",
-]
+requires-python = "==3.11.*" # Required by bpy 3.6.*
 
 [tool.uv]
 dev-dependencies = [
@@ -30,109 +26,117 @@ dev-dependencies = [
   "types-pygit2 ==1.15.0.20250319",
   "types-tqdm ==4.67.0.20250809",
   "vncdotool ==1.2.0",
+    "codespell ==2.4.1",
+    "dulwich ==0.22.7",
+    "mypy ==1.15.0",
+    "ruff ==0.9.6",
+    "starry-bpy-typings",
+    "tqdm ==4.67.1",
+    "types-tqdm ==4.67.0.20241221",
 ]
+environments = ["sys_platform == 'linux'"]
 
 [tool.uv.sources]
-bpy = {index = "download.blender.org"}
 starry-bpy-typings = {path = "typings"}
 
-[[tool.uv.index]]
-name = "download.blender.org"
-url = "https://download.blender.org/pypi/"
-
 [tool.ruff]
-target-version = "py39" # It overrides 'project.requires-python = "== 3.10.*"'
-src = ["typings/src", "src"]
+target-version = "py311" # It overrides 'project.requires-python = "== 3.11.*"'
+src = ["typings", "src"]
 
 [tool.ruff.lint]
 allowed-confusables = [
-  "\u30ce", # Katakana Letter No
+    "\u30ce", # Katakana Letter No
 ]
 select = ["ALL"]
 ignore = [
-  # ----- fixing -----
+    # ----- fixing -----
 
-  "FBT003", # Boolean positional value in function call
+    "FBT003", # Boolean positional value in function call
 
-  # ----- ignored -----
+    # ----- ignored -----
 
-  # `COM812` may cause conflicts when used with the formatter in Ruff 0.1.6
-  "COM812", # Trailing comma missing
+    # `COM812` may cause conflicts when used with the formatter in Ruff 0.1.6
+    "COM812", # Trailing comma missing
 
-  "D100", # Missing docstring in public module
-  "D101", # Missing docstring in public class
-  "D102", # Missing docstring in public method
-  "D103", # Missing docstring in public function
-  "D104", # Missing docstring in public package
-  "D105", # Missing docstring in magic method
-  "D106", # Missing docstring in public nested class
-  "D107", # Missing docstring in `__init__`
-  "ERA001", # Found commented-out code
+    "D100", # Missing docstring in public module
+    "D101", # Missing docstring in public class
+    "D102", # Missing docstring in public method
+    "D103", # Missing docstring in public function
+    "D104", # Missing docstring in public package
+    "D105", # Missing docstring in magic method
+    "D106", # Missing docstring in public nested class
+    "D107", # Missing docstring in `__init__`
+    "ERA001", # Found commented-out code
 
-  # devcontainer for windows hosts doesn't set executable permissions properly.
-  "EXE002", # The file is executable but no shebang is present
+    # devcontainer for windows hosts doesn't set executable permissions properly.
+    "EXE002", # The file is executable but no shebang is present
 
-  # https://projects.blender.org/blender/blender/issues/63713
-  "FA100", # Missing `from __future__ import annotations`, but uses `x`
+    # https://projects.blender.org/blender/blender/issues/63713
+    "FA100", # Missing `from __future__ import annotations`, but uses `x`
 
-  "FIX002", # Line contains TODO, consider resolving the issue
+    "FIX002", # Line contains TODO, consider resolving the issue
 
-  # `ISC001` may cause conflicts when used with the formatter in Ruff 0.1.6
-  "ISC001", # Implicitly concatenated string literals on one line
+    # `ISC001` may cause conflicts when used with the formatter in Ruff 0.1.6
+    "ISC001", # Implicitly concatenated string literals on one line
 
-  "ISC003", # Explicitly concatenated string should be implicitly concatenated
-  "PERF203", # `try`-`except` within a loop incurs performance overhead
-  "PLC0415", # `import` should be at the top-level of a file
-  "PLR2004", # Magic value used in comparison, consider replacing 3 with a constant variable
+    "ISC003", # Explicitly concatenated string should be implicitly concatenated
+    "PERF203", # `try`-`except` within a loop incurs performance overhead
+    "PLR2004", # Magic value used in comparison, consider replacing 3 with a constant variable
 
-  # pytest is not currently used.
-  "PT", # flake8-pytest-style
+    # pytest is not currently used.
+    "PT", # flake8-pytest-style
 
-  "RET504", # Unnecessary assignment to `x` before `return` statement
+    "RET504", # Unnecessary assignment to `x` before `return` statement
 
-  # SIM118 does not handle `bpy_prop_collection` correctly.
-  "SIM118", # Use `key not in dict` instead of `key not in dict.keys()`
+    # SIM118 does not handle `bpy_prop_collection` correctly.
+    "SIM118", # Use `key not in dict` instead of `key not in dict.keys()`
 
-  "TC001", # Move application import `x` into a type-checking block
-  "TC002", # Move third-party import `x` into a type-checking block
-  "TC003", # Move standard library import `x` into a type-checking block
-  "TD002", # Missing author in TODO; try: `# TODO(<author_name>): ...` or `# TODO @<author_name>: ...`
-  "TD003", # Missing issue link on the line following this TODO
+    "TC001", # Move application import `x` into a type-checking block
+    "TC002", # Move third-party import `x` into a type-checking block
+    "TC003", # Move standard library import `x` into a type-checking block
+    "TD002", # Missing author in TODO; try: `# TODO(<author_name>): ...` or `# TODO @<author_name>: ...`
+    "TD003", # Missing issue link on the line following this TODO
 
-  # https://projects.blender.org/blender/blender-manual/src/commit/585cc1a5eab811052ea4dda176dfce057cb6ad69/manual/advanced/extensions/addons.rst?display=source#L174-L183
-  "TID252", # Relative imports from parent modules are banned
+    # https://docs.blender.org/manual/en/4.2/advanced/extensions/addons.html#relative-imports
+    "TID252", # Relative imports from parent modules are banned
 ]
 
 [tool.ruff.lint.pep8-naming]
 extend-ignore-names = [
-  "glTF2ExportUserExtension",
-  "glTF2ImportUserExtension",
-  "EXPORT_SCENE_OT_*",
-  "IMPORT_SCENE_OT_*",
-  "ICYP_OT_*",
-  "VRM_FH_*",
-  "VRM_MT_*",
-  "VRM_OT_*",
-  "VRM_PT_*",
-  "VRM_UL_*",
-  "WM_OT_*",
+    "glTF2ExportUserExtension",
+    "glTF2ImportUserExtension",
+    "EXPORT_SCENE_OT_*",
+    "IMPORT_SCENE_OT_*",
+    "ICYP_OT_*",
+    "VRM_FH_*",
+    "VRM_OT_*",
+    "VRM_PT_*",
+    "VRM_UL_*",
+    "WM_OT_*",
 ]
 
 [tool.ruff.lint.pydocstyle]
 convention = "pep257"
 
 [tool.ruff.lint.mccabe]
-max-complexity = 216
+max-complexity = 150
 
 [tool.ruff.lint.pylint]
 max-args = 19
-max-branches = 215
+max-branches = 120
 max-returns = 25
 max-statements = 550
 
+[tool.codespell]
+ignore-regex = ".* # codespell-ignore"
+
 [tool.mypy]
-exclude = ['.']
-ignore_errors = true # Please use pyright instead.
+python_version = "3.11"
+mypy_path = "$MYPY_CONFIG_FILE_DIR/typings:$MYPY_CONFIG_FILE_DIR/src"
+strict = true
+disallow_any_unimported = true
+disallow_any_decorated = true
+disallow_any_explicit = true
 
 [tool.black]
 extend-exclude = ".*" # Please use Ruff instead.
