@@ -2,10 +2,9 @@
 """A module provides a function to convert a variable of type Any to a concrete type.
 
 Inevitably, variables of type Any may occur, and such variables cannot be handled
-in type checkers in strict mode. Any is allowed only in the module here.
+in mypy or pyright in strict mode. Any is allowed only in the module here.
 """
 
-import sys
 from collections.abc import Iterator
 from typing import (
     Any,  # Any is allowed only in the module here.
@@ -19,11 +18,7 @@ def to_object(  # type: ignore[explicit-any]
     # Interpret Unknown as object
     # https://github.com/microsoft/pyright/issues/3650
     if not isinstance(any_object, object):
-        if sys.version_info >= (3, 11):
-            from typing import assert_never
-
-            assert_never(any_object)
-        raise TypeError
+        raise TypeError  # typing.assert_never()
     return any_object
 
 
