@@ -13,11 +13,11 @@ def add_vrm0_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -27,12 +27,12 @@ def remove_vrm0_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     mesh_annotation_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         mesh_annotation_index=mesh_annotation_index,
     )
 
@@ -43,12 +43,12 @@ def move_up_vrm0_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     mesh_annotation_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         mesh_annotation_index=mesh_annotation_index,
     )
 
@@ -59,12 +59,12 @@ def move_down_vrm0_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     mesh_annotation_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         mesh_annotation_index=mesh_annotation_index,
     )
 
@@ -75,12 +75,12 @@ def add_vrm0_material_value_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_material_value_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
     )
 
@@ -91,13 +91,13 @@ def remove_vrm0_material_value_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     material_value_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_material_value_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         material_value_index=material_value_index,
     )
@@ -109,13 +109,13 @@ def move_up_vrm0_material_value_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     material_value_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_material_value_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         material_value_index=material_value_index,
     )
@@ -127,13 +127,13 @@ def move_down_vrm0_material_value_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     material_value_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_material_value_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         material_value_index=material_value_index,
     )
@@ -145,13 +145,13 @@ def add_vrm0_material_value_bind_target_value(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     material_value_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_material_value_bind_target_value(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         material_value_index=material_value_index,
     )
@@ -163,14 +163,14 @@ def remove_vrm0_material_value_bind_target_value(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     material_value_index: int = 0,
     target_value_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_material_value_bind_target_value(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         material_value_index=material_value_index,
         target_value_index=target_value_index,
@@ -183,12 +183,12 @@ def add_vrm0_blend_shape_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_blend_shape_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         name=name,
     )
 
@@ -199,12 +199,12 @@ def remove_vrm0_blend_shape_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_blend_shape_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
     )
 
@@ -215,12 +215,12 @@ def move_up_vrm0_blend_shape_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_blend_shape_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
     )
 
@@ -231,12 +231,12 @@ def move_down_vrm0_blend_shape_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_blend_shape_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
     )
 
@@ -247,12 +247,12 @@ def add_vrm0_blend_shape_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_blend_shape_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
     )
 
@@ -263,13 +263,13 @@ def remove_vrm0_blend_shape_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_blend_shape_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         bind_index=bind_index,
     )
@@ -281,13 +281,13 @@ def move_down_vrm0_blend_shape_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_blend_shape_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         bind_index=bind_index,
     )
@@ -299,29 +299,15 @@ def move_up_vrm0_blend_shape_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_blend_shape_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
         bind_index=bind_index,
-    )
-
-
-# This code is auto generated.
-# To regenerate, run the `uv run tools/property_typing.py` command.
-def restore_vrm0_blend_shape_group_bind_object(
-    execution_context: str = "EXEC_DEFAULT",
-    /,
-    *,
-    armature_object_name: str = "",
-) -> set[str]:
-    return bpy.ops.vrm.restore_vrm0_blend_shape_group_bind_object(  # type: ignore[attr-defined, no-any-return]
-        execution_context,
-        armature_object_name=armature_object_name,
     )
 
 
@@ -331,13 +317,13 @@ def add_vrm0_secondary_animation_collider_group_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
     bone_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_secondary_animation_collider_group_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
         bone_name=bone_name,
     )
@@ -349,13 +335,13 @@ def remove_vrm0_secondary_animation_collider_group_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_secondary_animation_collider_group_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
         collider_index=collider_index,
     )
@@ -367,13 +353,13 @@ def move_up_vrm0_secondary_animation_collider_group_coll(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_secondary_animation_collider_group_coll(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
         collider_index=collider_index,
     )
@@ -385,13 +371,13 @@ def move_down_vrm0_secondary_animation_collider_group_coll(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_secondary_animation_collider_group_coll(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
         collider_index=collider_index,
     )
@@ -403,12 +389,12 @@ def add_vrm0_secondary_animation_group_bone(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_secondary_animation_group_bone(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
     )
 
@@ -419,13 +405,13 @@ def remove_vrm0_secondary_animation_group_bone(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
     bone_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_secondary_animation_group_bone(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
         bone_index=bone_index,
     )
@@ -437,13 +423,13 @@ def move_up_vrm0_secondary_animation_group_bone(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
     bone_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_secondary_animation_group_bone(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
         bone_index=bone_index,
     )
@@ -455,13 +441,13 @@ def move_down_vrm0_secondary_animation_group_bone(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
     bone_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_secondary_animation_group_bone(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
         bone_index=bone_index,
     )
@@ -473,12 +459,12 @@ def add_vrm0_secondary_animation_group_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_secondary_animation_group_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
     )
 
@@ -489,13 +475,13 @@ def remove_vrm0_secondary_animation_group_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_secondary_animation_group_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
         collider_group_index=collider_group_index,
     )
@@ -507,13 +493,13 @@ def move_up_vrm0_secondary_animation_group_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_secondary_animation_group_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
         collider_group_index=collider_group_index,
     )
@@ -525,13 +511,13 @@ def move_down_vrm0_secondary_animation_group_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_secondary_animation_group_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
         collider_group_index=collider_group_index,
     )
@@ -543,12 +529,12 @@ def add_vrm0_secondary_animation_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     blend_shape_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_secondary_animation_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         blend_shape_group_index=blend_shape_group_index,
     )
 
@@ -559,12 +545,12 @@ def remove_vrm0_secondary_animation_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_secondary_animation_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
     )
 
@@ -575,12 +561,12 @@ def move_up_vrm0_secondary_animation_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_secondary_animation_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
     )
 
@@ -591,12 +577,12 @@ def move_down_vrm0_secondary_animation_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     bone_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_secondary_animation_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         bone_group_index=bone_group_index,
     )
 
@@ -607,11 +593,11 @@ def add_vrm0_secondary_animation_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm0_secondary_animation_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -621,12 +607,12 @@ def remove_vrm0_secondary_animation_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm0_secondary_animation_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
     )
 
@@ -637,12 +623,12 @@ def move_up_vrm0_secondary_animation_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm0_secondary_animation_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
     )
 
@@ -653,12 +639,12 @@ def move_down_vrm0_secondary_animation_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm0_secondary_animation_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
     )
 
@@ -669,27 +655,11 @@ def assign_vrm0_humanoid_human_bones_automatically(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.assign_vrm0_humanoid_human_bones_automatically(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
-    )
-
-
-# This code is auto generated.
-# To regenerate, run the `uv run tools/property_typing.py` command.
-def show_vrm0_bone_assignment_diagnostics(
-    execution_context: str = "EXEC_DEFAULT",
-    /,
-    *,
-    armature_object_name: str = "",
-    human_bone_name: str = "",
-) -> set[str]:
-    return bpy.ops.vrm.show_vrm0_bone_assignment_diagnostics(  # type: ignore[attr-defined, no-any-return]
-        execution_context,
-        armature_object_name=armature_object_name,
-        human_bone_name=human_bone_name,
+        armature_name=armature_name,
     )
 
 
@@ -699,11 +669,11 @@ def add_vrm1_meta_author(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm1_meta_author(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -713,12 +683,12 @@ def remove_vrm1_meta_author(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     author_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm1_meta_author(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         author_index=author_index,
     )
 
@@ -729,12 +699,12 @@ def move_up_vrm1_meta_author(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     author_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm1_meta_author(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         author_index=author_index,
     )
 
@@ -745,12 +715,12 @@ def move_down_vrm1_meta_author(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     author_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm1_meta_author(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         author_index=author_index,
     )
 
@@ -761,11 +731,11 @@ def add_vrm1_meta_reference(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm1_meta_reference(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -775,12 +745,12 @@ def remove_vrm1_meta_reference(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     reference_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm1_meta_reference(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         reference_index=reference_index,
     )
 
@@ -791,12 +761,12 @@ def move_up_vrm1_meta_reference(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     reference_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm1_meta_reference(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         reference_index=reference_index,
     )
 
@@ -807,12 +777,12 @@ def move_down_vrm1_meta_reference(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     reference_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm1_meta_reference(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         reference_index=reference_index,
     )
 
@@ -823,12 +793,12 @@ def add_vrm1_expressions_custom_expression(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     custom_expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm1_expressions_custom_expression(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         custom_expression_name=custom_expression_name,
     )
 
@@ -839,12 +809,12 @@ def remove_vrm1_expressions_custom_expression(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     custom_expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm1_expressions_custom_expression(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         custom_expression_name=custom_expression_name,
     )
 
@@ -855,12 +825,12 @@ def move_up_vrm1_expressions_custom_expression(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     custom_expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm1_expressions_custom_expression(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         custom_expression_name=custom_expression_name,
     )
 
@@ -871,12 +841,12 @@ def move_down_vrm1_expressions_custom_expression(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     custom_expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm1_expressions_custom_expression(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         custom_expression_name=custom_expression_name,
     )
 
@@ -887,12 +857,12 @@ def add_vrm1_expression_morph_target_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm1_expression_morph_target_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
     )
 
@@ -903,13 +873,13 @@ def remove_vrm1_expression_morph_target_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm1_expression_morph_target_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -921,13 +891,13 @@ def move_up_vrm1_expression_morph_target_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm1_expression_morph_target_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -939,29 +909,15 @@ def move_down_vrm1_expression_morph_target_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm1_expression_morph_target_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
-    )
-
-
-# This code is auto generated.
-# To regenerate, run the `uv run tools/property_typing.py` command.
-def restore_vrm1_expression_morph_target_bind_object(
-    execution_context: str = "EXEC_DEFAULT",
-    /,
-    *,
-    armature_object_name: str = "",
-) -> set[str]:
-    return bpy.ops.vrm.restore_vrm1_expression_morph_target_bind_object(  # type: ignore[attr-defined, no-any-return]
-        execution_context,
-        armature_object_name=armature_object_name,
     )
 
 
@@ -971,12 +927,12 @@ def add_vrm1_expression_material_color_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm1_expression_material_color_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
     )
 
@@ -987,13 +943,13 @@ def remove_vrm1_expression_material_color_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm1_expression_material_color_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -1005,13 +961,13 @@ def move_up_vrm1_expression_material_color_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm1_expression_material_color_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -1023,13 +979,13 @@ def move_down_vrm1_expression_material_color_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm1_expression_material_color_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -1041,12 +997,12 @@ def add_vrm1_expression_texture_transform_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm1_expression_texture_transform_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
     )
 
@@ -1057,13 +1013,13 @@ def remove_vrm1_expression_texture_transform_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm1_expression_texture_transform_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -1075,13 +1031,13 @@ def move_up_vrm1_expression_texture_transform_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm1_expression_texture_transform_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -1093,13 +1049,13 @@ def move_down_vrm1_expression_texture_transform_bind(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
     bind_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm1_expression_texture_transform_bind(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
         bind_index=bind_index,
     )
@@ -1111,11 +1067,11 @@ def add_vrm1_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_vrm1_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1125,12 +1081,12 @@ def remove_vrm1_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     mesh_annotation_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_vrm1_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         mesh_annotation_index=mesh_annotation_index,
     )
 
@@ -1141,12 +1097,12 @@ def move_up_vrm1_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     mesh_annotation_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_vrm1_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         mesh_annotation_index=mesh_annotation_index,
     )
 
@@ -1157,12 +1113,12 @@ def move_down_vrm1_first_person_mesh_annotation(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     mesh_annotation_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_vrm1_first_person_mesh_annotation(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         mesh_annotation_index=mesh_annotation_index,
     )
 
@@ -1173,11 +1129,11 @@ def assign_vrm1_humanoid_human_bones_automatically(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.assign_vrm1_humanoid_human_bones_automatically(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1197,29 +1153,13 @@ def refresh_vrm1_expression_texture_transform_bind_preview(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     expression_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.refresh_vrm1_expression_texture_transform_bind_preview(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         expression_name=expression_name,
-    )
-
-
-# This code is auto generated.
-# To regenerate, run the `uv run tools/property_typing.py` command.
-def show_vrm1_bone_assignment_diagnostics(
-    execution_context: str = "EXEC_DEFAULT",
-    /,
-    *,
-    armature_object_name: str = "",
-    human_bone_name: str = "",
-) -> set[str]:
-    return bpy.ops.vrm.show_vrm1_bone_assignment_diagnostics(  # type: ignore[attr-defined, no-any-return]
-        execution_context,
-        armature_object_name=armature_object_name,
-        human_bone_name=human_bone_name,
     )
 
 
@@ -1229,11 +1169,11 @@ def add_spring_bone1_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_spring_bone1_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1243,12 +1183,12 @@ def remove_spring_bone1_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_spring_bone1_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_index=collider_index,
     )
 
@@ -1259,12 +1199,12 @@ def move_up_spring_bone1_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_spring_bone1_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_index=collider_index,
     )
 
@@ -1275,12 +1215,12 @@ def move_down_spring_bone1_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_spring_bone1_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_index=collider_index,
     )
 
@@ -1291,11 +1231,11 @@ def add_spring_bone1_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_spring_bone1_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1305,12 +1245,12 @@ def remove_spring_bone1_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_spring_bone1_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
     )
 
@@ -1321,12 +1261,12 @@ def move_up_spring_bone1_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_spring_bone1_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
     )
 
@@ -1337,12 +1277,12 @@ def move_down_spring_bone1_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_spring_bone1_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
     )
 
@@ -1353,12 +1293,12 @@ def add_spring_bone1_collider_group_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_spring_bone1_collider_group_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
     )
 
@@ -1369,13 +1309,13 @@ def remove_spring_bone1_collider_group_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_spring_bone1_collider_group_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
         collider_index=collider_index,
     )
@@ -1387,13 +1327,13 @@ def move_up_spring_bone1_collider_group_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_spring_bone1_collider_group_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
         collider_index=collider_index,
     )
@@ -1405,13 +1345,13 @@ def move_down_spring_bone1_collider_group_collider(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     collider_group_index: int = 0,
     collider_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_spring_bone1_collider_group_collider(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         collider_group_index=collider_group_index,
         collider_index=collider_index,
     )
@@ -1423,11 +1363,11 @@ def add_spring_bone1_spring(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.add_spring_bone1_spring(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1437,12 +1377,12 @@ def remove_spring_bone1_spring(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_spring_bone1_spring(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
     )
 
@@ -1453,12 +1393,12 @@ def move_up_spring_bone1_spring(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_spring_bone1_spring(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
     )
 
@@ -1469,12 +1409,12 @@ def move_down_spring_bone1_spring(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_spring_bone1_spring(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
     )
 
@@ -1485,12 +1425,12 @@ def add_spring_bone1_spring_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.add_spring_bone1_spring_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
     )
 
@@ -1501,13 +1441,13 @@ def remove_spring_bone1_spring_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_spring_bone1_spring_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
         collider_group_index=collider_group_index,
     )
@@ -1519,13 +1459,13 @@ def move_up_spring_bone1_spring_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_spring_bone1_spring_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
         collider_group_index=collider_group_index,
     )
@@ -1537,13 +1477,13 @@ def move_down_spring_bone1_spring_collider_group(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
     collider_group_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_spring_bone1_spring_collider_group(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
         collider_group_index=collider_group_index,
     )
@@ -1555,13 +1495,13 @@ def add_spring_bone1_spring_joint(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
     guess_properties: bool = False,
 ) -> set[str]:
     return bpy.ops.vrm.add_spring_bone1_spring_joint(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
         guess_properties=guess_properties,
     )
@@ -1573,13 +1513,13 @@ def remove_spring_bone1_spring_joint(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
     joint_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.remove_spring_bone1_spring_joint(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
         joint_index=joint_index,
     )
@@ -1591,13 +1531,13 @@ def move_up_spring_bone1_joint(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
     joint_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_up_spring_bone1_joint(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
         joint_index=joint_index,
     )
@@ -1609,13 +1549,13 @@ def move_down_spring_bone1_joint(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
     spring_index: int = 0,
     joint_index: int = 0,
 ) -> set[str]:
     return bpy.ops.vrm.move_down_spring_bone1_joint(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
         spring_index=spring_index,
         joint_index=joint_index,
     )
@@ -1627,11 +1567,11 @@ def reset_spring_bone1_animation_state(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.reset_spring_bone1_animation_state(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1748,11 +1688,11 @@ def bones_rename(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.bones_rename(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1842,11 +1782,11 @@ def make_estimated_humanoid_t_pose(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    armature_object_name: str = "",
+    armature_name: str = "",
 ) -> set[str]:
     return bpy.ops.vrm.make_estimated_humanoid_t_pose(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        armature_object_name=armature_object_name,
+        armature_name=armature_name,
     )
 
 
@@ -1870,21 +1810,27 @@ def model_validate(
 
 # This code is auto generated.
 # To regenerate, run the `uv run tools/property_typing.py` command.
-def save_error_dialog_message(
+def import_vrm_via_file_handler(
     execution_context: str = "EXEC_DEFAULT",
     /,
     *,
-    filter_glob: str = "*.txt",
-    title: str = "",
-    lines: Optional[Sequence[Mapping[str, Union[str, int, float, bool]]]] = None,
     filepath: str = "",
-    check_existing: bool = True,
 ) -> set[str]:
-    return bpy.ops.vrm.save_error_dialog_message(  # type: ignore[attr-defined, no-any-return]
+    return bpy.ops.vrm.import_vrm_via_file_handler(  # type: ignore[attr-defined, no-any-return]
         execution_context,
-        filter_glob=filter_glob,
-        title=title,
-        lines=lines if lines is not None else [],
         filepath=filepath,
-        check_existing=check_existing,
+    )
+
+
+# This code is auto generated.
+# To regenerate, run the `uv run tools/property_typing.py` command.
+def import_vrma_via_file_handler(
+    execution_context: str = "EXEC_DEFAULT",
+    /,
+    *,
+    filepath: str = "",
+) -> set[str]:
+    return bpy.ops.vrm.import_vrma_via_file_handler(  # type: ignore[attr-defined, no-any-return]
+        execution_context,
+        filepath=filepath,
     )
