@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: MIT OR GPL-3.0-or-later
 from collections.abc import Sequence
-from typing import TYPE_CHECKING, Protocol, TypedDict, Union
+from typing import TYPE_CHECKING, Protocol, TypedDict
 
 from bpy.app.translations import pgettext
 from bpy.props import BoolProperty, IntVectorProperty
@@ -8,10 +8,6 @@ from bpy.types import AddonPreferences, Context, Operator, UILayout
 
 from . import version
 from .logger import get_logger
-from .shader import (
-    add_mtoon1_auto_setup_shader_node_group,
-    remove_mtoon1_auto_setup_shader_node_group,
-)
 
 logger = get_logger(__name__)
 
@@ -105,10 +101,11 @@ class ExportPreferencesProtocol(Protocol):
     export_invisibles: bool
     export_only_selections: bool
     enable_advanced_preferences: bool
+    export_ext_bmesh_encoding: bool
+    export_try_sparse_sk: bool
     export_all_influences: bool
     export_lights: bool
     export_gltf_animations: bool
-    export_try_sparse_sk: bool
 
 
 def copy_export_preferences(
@@ -118,55 +115,69 @@ def copy_export_preferences(
         destination.export_invisibles,
         destination.export_only_selections,
         destination.enable_advanced_preferences,
+        destination.export_ext_bmesh_encoding,
+        destination.export_try_sparse_sk,
         destination.export_all_influences,
         destination.export_lights,
         destination.export_gltf_animations,
-        destination.export_try_sparse_sk,
     ) = (
         source.export_invisibles,
         source.export_only_selections,
         source.enable_advanced_preferences,
+        source.export_ext_bmesh_encoding,
+        source.export_try_sparse_sk,
         source.export_all_influences,
         source.export_lights,
         source.export_gltf_animations,
-        source.export_try_sparse_sk,
     )
 
 
 def draw_advanced_options_description(
-    preferences: Union[AddonPreferences, Operator],
+    preferences: ExportPreferencesProtocol,
     property_name: str,
     layout: UILayout,
     description: str,
 ) -> None:
-    column = layout.box().column(align=True)
-    column.prop(preferences, property_name)
-    description_column = column.box().column(align=True)
-    for i, line in enumerate(description.splitlines()):
-        icon = "ERROR" if i == 0 else "NONE"
-        description_column.label(text=line, translate=False, icon=icon)
+    """Draw an advanced option with its description."""
+    row = layout.row()
+    row.prop(preferences, property_name)
+    
+    # Add description as help text
+    if description:
+        box = layout.box()
+        lines = description.split('\n')
+        for line in lines:
+            if line.strip():
+                box.label(text=line.strip())
 
 
 def draw_export_preferences_layout(
-    preferences: ExportPreferencesProtocol,
-    layout: UILayout,
-    *,
-    show_vrm1_options: bool,
+    preferences: ExportPreferencesProtocol, layout: UILayout
 ) -> None:
     if not isinstance(preferences, (AddonPreferences, Operator)):
         return
 
     layout.prop(preferences, "export_invisibles")
     layout.prop(preferences, "export_only_selections")
-
-    if not show_vrm1_options:
-        return
-
     layout.prop(preferences, "enable_advanced_preferences")
+    
     if not preferences.enable_advanced_preferences:
         return
 
     advanced_options_column = layout.box().column()
+
+    # EXT_bmesh_encoding extension
+    draw_advanced_options_description(
+        preferences,
+        "export_ext_bmesh_encoding",
+        advanced_options_column,
+        pgettext(
+            "Preserves complete BMesh topology\n"
+            + "(vertices, edges, loops, faces) using\n"
+            + "EXT_bmesh_encoding extension.\n"
+            + "Provides graceful fallback to standard glTF."
+        ),
+    )
 
     # UniVRM 0.115.0 doesn't support `export_try_sparse_sk`
     # https://github.com/saturday06/VRM-Addon-for-Blender/issues/381#issuecomment-1838365762
@@ -231,18 +242,6 @@ class VrmAddonPreferences(AddonPreferences):
         default=INITIAL_ADDON_VERSION,
     )
 
-    def update_add_mtoon_shader_node_group(self, context: Context) -> None:
-        if self.add_mtoon_shader_node_group:
-            add_mtoon1_auto_setup_shader_node_group(context)
-        else:
-            remove_mtoon1_auto_setup_shader_node_group(context)
-
-    add_mtoon_shader_node_group: BoolProperty(  # type: ignore[valid-type]
-        name="Add MToon shader node group",
-        default=True,
-        update=update_add_mtoon_shader_node_group,
-    )
-
     extract_textures_into_folder: BoolProperty(  # type: ignore[valid-type]
         name="Extract texture images into the folder",
         default=False,
@@ -286,17 +285,29 @@ class VrmAddonPreferences(AddonPreferences):
     enable_advanced_preferences: BoolProperty(  # type: ignore[valid-type]
         name="Enable Advanced Options",
     )
+    export_ext_bmesh_encoding: BoolProperty(  # type: ignore[valid-type]
+        name="Export EXT_bmesh_encoding",
+        description="Enable BMesh topology preservation using EXT_bmesh_encoding extension",
+        default=False,
+    )
+    export_try_sparse_sk: BoolProperty(  # type: ignore[valid-type]
+        name="Export Sparse Shape Keys",
+        description="Reduces file size but may not be readable by older apps",
+        default=False,
+    )
     export_all_influences: BoolProperty(  # type: ignore[valid-type]
         name="Export All Bone Influences",
+        description="Don't limit to 4, most viewers truncate to 4, "
+        + "so bone movement may cause jagged meshes",
+        # The upstream says that Models may appear incorrectly in many viewers.
+        # https://github.com/KhronosGroup/glTF-Blender-IO/blob/356b3dda976303d3ecce8b3bd1591245e576db38/addons/io_scene_gltf2/__init__.py#L760
+        default=False,
     )
     export_lights: BoolProperty(  # type: ignore[valid-type]
         name="Export Lights",
     )
     export_gltf_animations: BoolProperty(  # type: ignore[valid-type]
         name="Export glTF Animations",
-    )
-    export_try_sparse_sk: BoolProperty(  # type: ignore[valid-type]
-        name="Use Sparse Accessors",
     )
 
     def draw(self, _context: Context) -> None:
@@ -313,21 +324,18 @@ class VrmAddonPreferences(AddonPreferences):
                     icon="NONE" if index else "ERROR",
                 )
 
-        layout.prop(self, "add_mtoon_shader_node_group")
-
         import_box = layout.box()
         import_box.label(text="Import", icon="IMPORT")
         draw_import_preferences_layout(self, import_box)
 
         export_box = layout.box()
         export_box.label(text="Export", icon="EXPORT")
-        draw_export_preferences_layout(self, export_box, show_vrm1_options=True)
+        draw_export_preferences_layout(self, export_box)
 
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
         addon_version: Sequence[int]  # type: ignore[no-redef]
-        add_mtoon_shader_node_group: bool  # type: ignore[no-redef]
         extract_textures_into_folder: bool  # type: ignore[no-redef]
         make_new_texture_folder: bool  # type: ignore[no-redef]
         set_shading_type_to_material_on_import: bool  # type: ignore[no-redef]
@@ -339,10 +347,11 @@ class VrmAddonPreferences(AddonPreferences):
         export_invisibles: bool  # type: ignore[no-redef]
         export_only_selections: bool  # type: ignore[no-redef]
         enable_advanced_preferences: bool  # type: ignore[no-redef]
+        export_ext_bmesh_encoding: bool  # type: ignore[no-redef]
+        export_try_sparse_sk: bool  # type: ignore[no-redef]
         export_all_influences: bool  # type: ignore[no-redef]
         export_lights: bool  # type: ignore[no-redef]
         export_gltf_animations: bool  # type: ignore[no-redef]
-        export_try_sparse_sk: bool  # type: ignore[no-redef]
 
 
 def get_preferences(context: Context) -> VrmAddonPreferences:
