@@ -55,12 +55,10 @@ class Progress:
         if not self.show_progress:
             return
 
-        # The mouse cursor becomes a four-digit number, and there is an area
-        # where values from 0 to 9999 can be displayed. However, if the
-        # progress rate is directly converted to a number from 0 to 9999, the
-        # last two digits will frequently round-trip, making the progress
-        # difficult to understand. Therefore, only the last two digits display
-        # area is used to show the progress rate as a number from 0 to 99.
+        # マウスカーソルが四桁の数値になり、0から9999までの値が表示できる領域がある
+        # しかし、進捗率をそのまま0から9999の数値に変換すると、下二桁の数値が頻繁に
+        # ラウンドトリップし進捗状況が分かりにくくなる。そのため、下二桁の表示領域
+        # だけを利用し0から99の数値で進捗率を表示する
         self.context.window_manager.progress_update(math.floor(ratio * 99))
 
 
