@@ -16,9 +16,14 @@ from .extension import (
     get_armature_extension,
 )
 from .mtoon1 import migration as mtoon1_migration
+from .mtoon1 import ops as mtoon1_ops
+from .property_group import BonePropertyGroup
+from .bmesh_encoding import migration as bmesh_encoding_migration
 from .spring_bone1 import migration as spring_bone1_migration
 from .vrm0 import migration as vrm0_migration
+from .vrm0.property_group import Vrm0HumanoidPropertyGroup
 from .vrm1 import migration as vrm1_migration
+from .vrm1 import property_group as vrm1_property_group
 
 logger = get_logger(__name__)
 
@@ -34,9 +39,11 @@ state: Final = State()
 
 def is_unnecessary(armature_data: Armature) -> bool:
     ext = get_armature_extension(armature_data)
-    return tuple(
-        ext.addon_version
-    ) >= get_addon_version() and vrm0_migration.is_unnecessary(ext.vrm0)
+    return (
+        tuple(ext.addon_version) >= get_addon_version()
+        and armature_data.name == ext.armature_data_name
+        and vrm0_migration.is_unnecessary(ext.vrm0)
+    )
 
 
 def defer_migrate(armature_object_name: str) -> bool:
@@ -60,8 +67,8 @@ def defer_migrate(armature_object_name: str) -> bool:
 
 
 def migrate_timer_callback(armature_object_name: str) -> None:
-    """Match the type of migrate() to bpy.app.timers.register."""
-    context = bpy.context  # Context cannot span frames, so get it anew
+    """migrate()の型をbpy.app.timers.registerに合わせるためのラッパー."""
+    context = bpy.context  # Contextはフレームを跨げないので新たに取得する
     migrate(context, armature_object_name)
 
 
@@ -80,10 +87,15 @@ def migrate(context: Optional[Context], armature_object_name: str) -> bool:
         return True
 
     ext = get_armature_extension(armature_data)
+    ext.armature_data_name = armature_data.name
+
+    for bone_property_group in BonePropertyGroup.get_all_bone_property_groups(armature):
+        bone_property_group.armature_data_name = armature_data.name
 
     vrm0_migration.migrate(context, ext.vrm0, armature)
     vrm1_migration.migrate(context, ext.vrm1, armature)
     spring_bone1_migration.migrate(context, armature)
+    bmesh_encoding_migration.migrate(context, armature)
 
     updated_addon_version = get_addon_version()
     logger.info(
@@ -103,9 +115,11 @@ def migrate_all_objects(
     show_progress: bool = False,
 ) -> None:
     for obj in context.blend_data.objects:
-        if obj.type == "ARMATURE" and isinstance(armature_data := obj.data, Armature):
+        if obj.type == "ARMATURE":
             if skip_non_migrated_armatures:
-                ext = get_armature_extension(armature_data)
+                ext = getattr(obj.data, "vrm_addon_extension", None)
+                if not isinstance(ext, VrmAddonArmatureExtensionPropertyGroup):
+                    continue
                 if (
                     tuple(ext.addon_version)
                     == VrmAddonArmatureExtensionPropertyGroup.INITIAL_ADDON_VERSION
@@ -129,9 +143,7 @@ def migrate_all_objects(
         updated_addon_version,
     )
 
-    if tuple(preferences.addon_version) != preferences.INITIAL_ADDON_VERSION and tuple(
-        preferences.addon_version
-    ) < (2, 34, 0):
+    if tuple(preferences.addon_version) < (2, 34, 0):
         preferences.enable_advanced_preferences = True
         preferences.export_gltf_animations = True
 
@@ -139,11 +151,10 @@ def migrate_all_objects(
 
 
 def validate_blend_file_compatibility(context: Context) -> None:
-    """Warn when attempting to edit a file created in newer Blender with older Blender.
+    """新しいBlenderで作成されたファイルを古いBlenderで編集しようとした場合に警告をする.
 
-    Due to add-on version support issues, there are often reports of users trying to
-    edit files created in newer Blender with older Blender, which can cause shape keys
-    to break. This warning alerts users to be careful.
+    アドオンの対応バージョンの事情で新しいBlenderで編集されたファイルを古いBlenderで編集しようとし、
+    それによりシェイプキーが壊れるなどの報告がよく上がる。警告を出すことでユーザーに注意を促す。
     """
     if not context.blend_data.filepath:
         return
@@ -169,8 +180,7 @@ def validate_blend_file_compatibility(context: Context) -> None:
 
     if not state.blend_file_compatibility_warning_shown:
         state.blend_file_compatibility_warning_shown = True
-        # Use timer because dialog disappears automatically if not executed with
-        # timer in Blender 4.2.0
+        # Blender 4.2.0ではtimerで実行しないとダイアログが自動で消えるのでタイマーを使う
         bpy.app.timers.register(
             functools.partial(
                 show_blend_file_compatibility_warning,
@@ -190,15 +200,12 @@ def show_blend_file_compatibility_warning(file_version: str, app_version: str) -
 
 
 def validate_blend_file_addon_compatibility(context: Context) -> None:
-    """Warn when attempting to edit a file created with a newer VRM add-on.
-
-    This warning is for files using an older VRM add-on.
-    """
+    """新しいVRMアドオンで作成されたファイルを古いVRMアドオンで編集しようとした場合に警告をする."""
     if not context.blend_data.filepath:
         return
     installed_addon_version = get_addon_version()
 
-    # TODO: It might be better to store the version in Scene or similar
+    # TODO: これはSceneあたりにバージョンを生やしたほうが良いかも
     up_to_date = True
     file_addon_version: tuple[int, ...] = (0, 0, 0)
     for armature in context.blend_data.armatures:
@@ -220,8 +227,7 @@ def validate_blend_file_addon_compatibility(context: Context) -> None:
 
     if not state.blend_file_compatibility_warning_shown:
         state.blend_file_compatibility_warning_shown = True
-        # Use timer because dialog disappears automatically if not executed with
-        # timer in Blender 4.2.0
+        # Blender 4.2.0ではtimerで実行しないとダイアログが自動で消えるのでタイマーを使う
         bpy.app.timers.register(
             functools.partial(
                 show_blend_file_addon_compatibility_warning,
@@ -265,3 +271,60 @@ def have_vrm_model(context: Context) -> bool:
 
         return True
     return False
+
+
+def on_change_bpy_object_name() -> None:
+    context = bpy.context
+
+    for armature in context.blend_data.armatures:
+        ext = getattr(armature, "vrm_addon_extension", None)
+        if not isinstance(ext, VrmAddonArmatureExtensionPropertyGroup):
+            continue
+        if (
+            tuple(ext.addon_version)
+            == VrmAddonArmatureExtensionPropertyGroup.INITIAL_ADDON_VERSION
+        ):
+            continue
+
+        # TODO: Needs optimization!
+        for collider in ext.spring_bone1.colliders:
+            collider.broadcast_bpy_object_name(context)
+
+
+def on_change_bpy_object_mode() -> None:
+    context = bpy.context
+
+    active_object = context.active_object
+    if not active_object:
+        return
+    mtoon1_ops.VRM_OT_refresh_mtoon1_outline.refresh_object(context, active_object)
+
+
+def on_change_bpy_object_location() -> None:
+    context = bpy.context
+    active_object = context.active_object
+    if not active_object:
+        return
+    vrm1_property_group.Vrm1LookAtPropertyGroup.update_all_previews(context)
+
+
+def on_change_bpy_bone_name() -> None:
+    context = bpy.context
+
+    for armature in context.blend_data.armatures:
+        ext = getattr(armature, "vrm_addon_extension", None)
+        if not isinstance(ext, VrmAddonArmatureExtensionPropertyGroup):
+            continue
+        if (
+            tuple(ext.addon_version)
+            == VrmAddonArmatureExtensionPropertyGroup.INITIAL_ADDON_VERSION
+        ):
+            continue
+
+        Vrm0HumanoidPropertyGroup.update_all_node_candidates(context, armature.name)
+
+
+def on_change_bpy_armature_name() -> None:
+    context = bpy.context
+
+    migrate_all_objects(context, skip_non_migrated_armatures=True)
