@@ -1,8 +1,9 @@
 # SPDX-License-Identifier: MIT OR GPL-3.0-or-later
 # SPDX-FileCopyrightText: 2018 iCyP
 # SPDX-FileCopyrightText: 2024 saturday06
+
 import json
-import warnings
+import re
 import webbrowser
 from collections.abc import Set as AbstractSet
 from pathlib import Path
@@ -21,55 +22,35 @@ from bpy.types import (
 from bpy_extras.io_utils import ExportHelper, ImportHelper
 
 from ..common.deep import make_json
-from ..common.human_bone_mapper.vroid_mapping import (
-    FULL_PATTERN,
-    symmetrise_vroid_bone_name,
-)
-from ..common.logger import get_logger
 from ..common.vrm0.human_bone import HumanBoneSpecifications
 from ..common.workspace import save_workspace
 from . import search
 from .extension import get_armature_extension
 from .t_pose import set_estimated_humanoid_t_pose
 
-logger = get_logger(__name__)
-
 
 class VRM_OT_simplify_vroid_bones(Operator):
     bl_idname = "vrm.bones_rename"
     bl_label = "Symmetrize VRoid Bone Names on X-Axis"
-    bl_description = "Make VRoid bone names editable for X-axis mirroring."
+    bl_description = "Make VRoid bone names X-axis mirror editable"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
+    left__pattern = re.compile("^J_(Adj|Bip|Opt|Sec)_L_")
+    right_pattern = re.compile("^J_(Adj|Bip|Opt|Sec)_R_")
+    full__pattern = re.compile("^J_(Adj|Bip|Opt|Sec)_([CLR]_)?")
 
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     @staticmethod
     def vroid_bones_exist(armature: Armature) -> bool:
-        return any(map(FULL_PATTERN.match, armature.bones.keys()))
+        return any(
+            map(VRM_OT_simplify_vroid_bones.full__pattern.match, armature.bones.keys())
+        )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -78,21 +59,33 @@ class VRM_OT_simplify_vroid_bones(Operator):
 
         with save_workspace(context):
             for bone_name, bone in armature_data.bones.items():
-                bone.name = symmetrise_vroid_bone_name(bone_name)
+                left = VRM_OT_simplify_vroid_bones.left__pattern.sub("", bone_name)
+                if left != bone_name:
+                    bone.name = left + "_L"
+                    continue
+
+                right = VRM_OT_simplify_vroid_bones.right_pattern.sub("", bone_name)
+                if right != bone_name:
+                    bone.name = right + "_R"
+                    continue
+
+                a = VRM_OT_simplify_vroid_bones.full__pattern.sub("", bone_name)
+                if a != bone_name:
+                    bone.name = a
+                    continue
 
         return {"FINISHED"}
 
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
 
 
 class VRM_OT_save_human_bone_mappings(Operator, ExportHelper):
     bl_idname = "vrm.save_human_bone_mappings"
     bl_label = "Save Bone Mappings"
-    bl_description = "Save bone mappings."
+    bl_description = "Save Bone Mappings"
     bl_options: AbstractSet[str] = {"REGISTER"}
 
     filename_ext = ".json"
@@ -142,7 +135,7 @@ class VRM_OT_save_human_bone_mappings(Operator, ExportHelper):
 class VRM_OT_load_human_bone_mappings(Operator, ImportHelper):
     bl_idname = "vrm.load_human_bone_mappings"
     bl_label = "Load Bone Mappings"
-    bl_description = "Load bone mappings."
+    bl_description = "Load Bone Mappings"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
     filename_ext = ".json"
@@ -178,7 +171,7 @@ class VRM_OT_load_human_bone_mappings(Operator, ImportHelper):
                 armature_data
             ).vrm0.humanoid.human_bones:
                 if human_bone.bone == human_bone_name:
-                    human_bone.node.bone_name = bpy_bone_name
+                    human_bone.node.set_bone_name(bpy_bone_name)
                     found = True
                     break
             if found:
@@ -188,7 +181,7 @@ class VRM_OT_load_human_bone_mappings(Operator, ImportHelper):
                 armature_data
             ).vrm0.humanoid.human_bones.add()
             human_bone.bone = human_bone_name
-            human_bone.node.bone_name = bpy_bone_name
+            human_bone.node.set_bone_name(bpy_bone_name)
 
         return {"FINISHED"}
 
@@ -204,7 +197,7 @@ class VRM_OT_load_human_bone_mappings(Operator, ImportHelper):
 class VRM_OT_open_url_in_web_browser(Operator):
     bl_idname = "vrm.open_url_in_web_browser"
     bl_label = "Open"
-    bl_description = "Open the URL in the default web browser."
+    bl_description = "Open the URL in the default web browser"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
     url: StringProperty(  # type: ignore[valid-type]
@@ -235,7 +228,7 @@ class VRM_OT_open_url_in_web_browser(Operator):
 class VRM_OT_show_blend_file_compatibility_warning(Operator):
     bl_idname = "vrm.show_blend_file_compatibility_warning"
     bl_label = "File Compatibility Warning"
-    bl_description = "Show blend file compatibility warning."
+    bl_description = "Show Blend File Compatibility Warning"
     bl_options: AbstractSet[str] = {"REGISTER"}
 
     file_version: StringProperty(options={"HIDDEN"})  # type: ignore[valid-type]
@@ -253,7 +246,7 @@ class VRM_OT_show_blend_file_compatibility_warning(Operator):
             "The current file is not compatible with the running Blender.\n"
             + "The current file was created in Blender {file_version}, but the running"
             + " Blender version is {app_version}.\n"
-            + "This incompatibility may result in data loss or corruption."
+            + "So it is not compatible. As a result some data may be lost or corrupted."
         ).format(
             app_version=self.app_version,
             file_version=self.file_version,
@@ -282,7 +275,7 @@ class VRM_OT_show_blend_file_compatibility_warning(Operator):
 class VRM_OT_show_blend_file_addon_compatibility_warning(Operator):
     bl_idname = "vrm.show_blend_file_addon_compatibility_warning"
     bl_label = "VRM Add-on Compatibility Warning"
-    bl_description = "Show blend file and VRM add-on compatibility warning."
+    bl_description = "Show Blend File and VRM Add-on Compatibility Warning"
     bl_options: AbstractSet[str] = {"REGISTER"}
 
     file_addon_version: StringProperty(options={"HIDDEN"})  # type: ignore[valid-type]
@@ -300,8 +293,9 @@ class VRM_OT_show_blend_file_addon_compatibility_warning(Operator):
             "The current file is not compatible with the installed VRM Add-on.\n"
             + "The current file was created in VRM Add-on {file_addon_version}, but the"
             + " installed\n"
-            + "VRM Add-on version is {installed_addon_version}. This incompatibility\n"
-            + "may result in data loss or corruption."
+            + "VRM Add-on version is {installed_addon_version}. So it is not"
+            + " compatible. As a result some\n"
+            + "data may be lost or corrupted."
         ).format(
             file_addon_version=self.file_addon_version,
             installed_addon_version=self.installed_addon_version,
@@ -358,40 +352,21 @@ def layout_operator(
             f"{type(operator)} is not compatible with {operator_type}."
             + f"the expected name is {name}"
         )
-    return cast("__Operator", operator)
+    return cast(__Operator, operator)
 
 
 class VRM_OT_make_estimated_humanoid_t_pose(Operator):
     bl_idname = "vrm.make_estimated_humanoid_t_pose"
     bl_label = "Make Estimated T-Pose"
-    bl_description = "Create VRM estimated humanoid T-pose."
+    bl_description = "Make VRM Estimated Humanoid T-Pose"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
-
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
 
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         if not set_estimated_humanoid_t_pose(context, armature):
@@ -401,5 +376,4 @@ class VRM_OT_make_estimated_humanoid_t_pose(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
