@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: MIT OR GPL-3.0-or-later
 import statistics
+import uuid
 from collections.abc import Sequence
 from sys import float_info
 from typing import TYPE_CHECKING, Callable, Optional
@@ -18,7 +19,7 @@ from bpy.props import (
 from bpy.types import Armature, Context, Object, PropertyGroup
 from mathutils import Matrix, Quaternion, Vector
 
-from ...common import convert, safe_removal
+from ...common import convert
 from ...common.logger import get_logger
 from ...common.rotation import (
     get_rotation_as_quaternion,
@@ -658,15 +659,7 @@ class SpringBone1ColliderExtensionsPropertyGroup(PropertyGroup):
 
 # https://github.com/vrm-c/vrm-specification/blob/6fb6baaf9b9095a84fb82c8384db36e1afeb3558/specification/VRMC_springBone-1.0-beta/schema/VRMC_springBone.collider.schema.json
 class SpringBone1ColliderPropertyGroup(PropertyGroup):
-    def broadcast_bpy_object_name(self) -> None:
-        armature = self.id_data
-        if not isinstance(armature, Armature):
-            message = (
-                f"{type(self)}/{self}.id_data is not a {Armature}"
-                + f" but {type(armature)}/{armature}"
-            )
-            raise TypeError(message)
-
+    def broadcast_bpy_object_name(self, context: Context) -> None:
         if not self.bpy_object or not self.bpy_object.name:
             self.name = ""
             return
@@ -674,13 +667,24 @@ class SpringBone1ColliderPropertyGroup(PropertyGroup):
             return
         self.name = self.bpy_object.name
 
-        spring_bone = get_armature_spring_bone1_extension(armature)
-        for collider_group in spring_bone.collider_groups:
-            for collider_reference in collider_group.colliders:
-                if self.uuid != collider_reference.collider_uuid:
+        self.search_one_time_uuid = uuid.uuid4().hex
+        for armature in context.blend_data.armatures:
+            if not hasattr(armature, "vrm_addon_extension"):
+                continue
+
+            spring_bone = get_armature_spring_bone1_extension(armature)
+
+            for collider in spring_bone.colliders:
+                if collider.search_one_time_uuid != self.search_one_time_uuid:
                     continue
-                if collider_reference.collider_name != self.name:
-                    collider_reference.collider_name = self.name
+
+                for collider_group in spring_bone.collider_groups:
+                    for collider_reference in collider_group.colliders:
+                        if self.uuid != collider_reference.collider_uuid:
+                            continue
+                        collider_reference.collider_name = self.name
+
+                return
 
     node: PointerProperty(type=BonePropertyGroup)  # type: ignore[valid-type]
     shape: PointerProperty(type=SpringBone1ColliderShapePropertyGroup)  # type: ignore[valid-type]
@@ -805,6 +809,7 @@ class SpringBone1ColliderPropertyGroup(PropertyGroup):
 
     # for references
     uuid: StringProperty()  # type: ignore[valid-type]
+    search_one_time_uuid: StringProperty()  # type: ignore[valid-type]
 
     def reset_bpy_object(self, context: Context, armature: Object) -> None:
         armature_data = armature.data
@@ -838,14 +843,15 @@ class SpringBone1ColliderPropertyGroup(PropertyGroup):
                 extended.SHAPE_TYPE_EXTENDED_PLANE.identifier,
             ]
         ):
-            for child in list(self.bpy_object.children):
-                if not safe_removal.remove_object(context, child):
-                    logger.warning(
-                        'Failed to remove "%s" with %d users'
-                        " while changing spring bone collider type",
-                        child.name,
-                        child.users,
-                    )
+            children = list(self.bpy_object.children)
+            for collection in context.blend_data.collections:
+                for child in children:
+                    child.parent = None
+                    if child.name in collection.objects:
+                        collection.objects.unlink(child)
+            for child in children:
+                if child.users <= 1:
+                    context.blend_data.objects.remove(child, do_unlink=True)
             empty_display_type = "SPHERE"
             if extended.shape_type == extended.SHAPE_TYPE_EXTENDED_PLANE.identifier:
                 empty_display_type = "CIRCLE"
@@ -889,7 +895,7 @@ class SpringBone1ColliderPropertyGroup(PropertyGroup):
             if self.bpy_object.parent_bone:
                 self.bpy_object.parent_bone = ""
 
-        self.broadcast_bpy_object_name()
+        self.broadcast_bpy_object_name(context)
 
     if TYPE_CHECKING:
         # This code is auto generated.
@@ -902,6 +908,7 @@ class SpringBone1ColliderPropertyGroup(PropertyGroup):
         ui_collider_type: str  # type: ignore[no-redef]
         bpy_object: Optional[Object]  # type: ignore[no-redef]
         uuid: str  # type: ignore[no-redef]
+        search_one_time_uuid: str  # type: ignore[no-redef]
 
 
 class SpringBone1ColliderReferencePropertyGroup(PropertyGroup):
@@ -910,13 +917,7 @@ class SpringBone1ColliderReferencePropertyGroup(PropertyGroup):
         return value if isinstance(value, str) else str(value)
 
     def set_collider_name(self, value: object) -> None:
-        armature = self.id_data
-        if not isinstance(armature, Armature):
-            message = (
-                f"{type(self)}/{self}.id_data is not a {Armature}"
-                + f" but {type(armature)}/{armature}"
-            )
-            raise TypeError(message)
+        context = bpy.context
 
         if not isinstance(value, str):
             value = str(value)
@@ -925,23 +926,34 @@ class SpringBone1ColliderReferencePropertyGroup(PropertyGroup):
             return
         self["collider_name"] = value
 
-        spring_bone = get_armature_spring_bone1_extension(armature)
-        for collider in spring_bone.colliders:
-            if collider.name != value:
-                continue
-            if self.collider_uuid != collider.uuid:
-                self.collider_uuid = collider.uuid
+        self.search_one_time_uuid = uuid.uuid4().hex
+        for armature in context.blend_data.armatures:
+            spring_bone = get_armature_spring_bone1_extension(armature)
+            for collider_group in spring_bone.collider_groups:
+                for collider_reference in collider_group.colliders:
+                    if (
+                        collider_reference.search_one_time_uuid
+                        != self.search_one_time_uuid
+                    ):
+                        continue
+
+                    for collider in spring_bone.colliders:
+                        if collider.name == value:
+                            self.collider_uuid = collider.uuid
+                    return
 
     collider_name: StringProperty(  # type: ignore[valid-type]
         get=get_collider_name, set=set_collider_name
     )
     collider_uuid: StringProperty()  # type: ignore[valid-type]
+    search_one_time_uuid: StringProperty()  # type: ignore[valid-type]
 
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
         collider_name: str  # type: ignore[no-redef]
         collider_uuid: str  # type: ignore[no-redef]
+        search_one_time_uuid: str  # type: ignore[no-redef]
 
 
 # https://github.com/vrm-c/vrm-specification/blob/f2d8f158297fc883aef9c3071ca68fbe46b03f45/specification/0.0/schema/vrm.secondaryanimation.collidergroup.schema.json
@@ -951,35 +963,31 @@ class SpringBone1ColliderGroupPropertyGroup(PropertyGroup):
         return value if isinstance(value, str) else str(value)
 
     def set_vrm_name(self, vrm_name: object) -> None:
+        context = bpy.context
+
         if not isinstance(vrm_name, str):
             vrm_name = str(vrm_name)
         self["vrm_name"] = vrm_name
-        self.fix_index()
+        self.fix_index(context)
 
-    def fix_index(self) -> None:
-        armature = self.id_data
-        if not isinstance(armature, Armature):
-            message = (
-                f"{type(self)}/{self}.id_data is not a {Armature}"
-                + f" but {type(armature)}/{armature}"
-            )
-            raise TypeError(message)
+    def fix_index(self, context: Context) -> None:
+        self.search_one_time_uuid = uuid.uuid4().hex
+        for armature in context.blend_data.armatures:
+            spring_bone = get_armature_spring_bone1_extension(armature)
 
-        spring_bone = get_armature_spring_bone1_extension(armature)
+            for index, collider_group in enumerate(spring_bone.collider_groups):
+                if collider_group.search_one_time_uuid != self.search_one_time_uuid:
+                    continue
 
-        for index, collider_group in enumerate(spring_bone.collider_groups):
-            if collider_group.path_from_id() != self.path_from_id():
-                continue
+                name = f"{self.vrm_name} #{index + 1}"
+                self.name = name
 
-            name = f"{self.vrm_name} #{index + 1}"
-            self.name = name
+                for spring in spring_bone.springs:
+                    for collider_group_reference in spring.collider_groups:
+                        if collider_group_reference.collider_group_uuid == self.uuid:
+                            collider_group_reference.collider_group_name = name
 
-            for spring in spring_bone.springs:
-                for collider_group_reference in spring.collider_groups:
-                    if collider_group_reference.collider_group_uuid == self.uuid:
-                        collider_group_reference.collider_group_name = name
-
-            return
+                return
 
     vrm_name: StringProperty(  # type: ignore[valid-type]
         name="Name",
@@ -997,8 +1005,10 @@ class SpringBone1ColliderGroupPropertyGroup(PropertyGroup):
     active_collider_index: IntProperty(min=0)  # type: ignore[valid-type]
 
     # for reference
-    # When copying objects, there may be multiple objects with the same uuid
+    # オブジェクトをコピーした場合同じuuidをもつオブジェクトが複数ある可能性がある
     uuid: StringProperty()  # type: ignore[valid-type]
+
+    search_one_time_uuid: StringProperty()  # type: ignore[valid-type]
 
     if TYPE_CHECKING:
         # This code is auto generated.
@@ -1010,6 +1020,7 @@ class SpringBone1ColliderGroupPropertyGroup(PropertyGroup):
         show_expanded: bool  # type: ignore[no-redef]
         active_collider_index: int  # type: ignore[no-redef]
         uuid: str  # type: ignore[no-redef]
+        search_one_time_uuid: str  # type: ignore[no-redef]
 
 
 # https://github.com/vrm-c/vrm-specification/tree/993a90a5bda9025f3d9e2923ad6dea7506f88553/specification/VRMC_springBone-1.0#initialization
@@ -1105,38 +1116,44 @@ class SpringBone1ColliderGroupReferencePropertyGroup(PropertyGroup):
         return value if isinstance(value, str) else str(value)
 
     def set_collider_group_name(self, value: object) -> None:
-        armature = self.id_data
-        if not isinstance(armature, Armature):
-            message = (
-                f"{type(self)}/{self}.id_data is not a {Armature}"
-                + f" but {type(armature)}/{armature}"
-            )
-            raise TypeError(message)
+        context = bpy.context
 
         if not isinstance(value, str):
             value = str(value)
         self.name = value
-        if self.get("collider_group_name") != value:
-            self["collider_group_name"] = value
+        if self.get("collider_group_name") == value:
+            return
+        self["collider_group_name"] = value
 
-        spring_bone = get_armature_spring_bone1_extension(armature)
-        for collider_group in spring_bone.collider_groups:
-            if collider_group.name != value:
-                continue
-            if self.collider_group_uuid != collider_group.uuid:
-                self.collider_group_uuid = collider_group.uuid
-            break
+        self.search_one_time_uuid = uuid.uuid4().hex
+        for armature in context.blend_data.armatures:
+            spring_bone = get_armature_spring_bone1_extension(armature)
+            for spring in spring_bone.springs:
+                for collider_group_reference in spring.collider_groups:
+                    if (
+                        collider_group_reference.search_one_time_uuid
+                        != self.search_one_time_uuid
+                    ):
+                        continue
+
+                    for collider_group in spring_bone.collider_groups:
+                        if collider_group.name == value:
+                            self.collider_group_uuid = collider_group.uuid
+                            break
+                    return
 
     collider_group_name: StringProperty(  # type: ignore[valid-type]
         get=get_collider_group_name, set=set_collider_group_name
     )
     collider_group_uuid: StringProperty()  # type: ignore[valid-type]
+    search_one_time_uuid: StringProperty()  # type: ignore[valid-type]
 
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
         collider_group_name: str  # type: ignore[no-redef]
         collider_group_uuid: str  # type: ignore[no-redef]
+        search_one_time_uuid: str  # type: ignore[no-redef]
 
 
 class SpringBone1SpringAnimationStatePropertyGroup(PropertyGroup):
@@ -1264,8 +1281,9 @@ class SpringBone1SpringBonePropertyGroup(PropertyGroup):
 def get_armature_spring_bone1_extension(
     armature: Armature,
 ) -> SpringBone1SpringBonePropertyGroup:
-    from ..extension import get_armature_extension
-
-    ext = get_armature_extension(armature)
-    spring_bone1: SpringBone1SpringBonePropertyGroup = ext.spring_bone1
+    spring_bone1 = getattr(
+        getattr(armature, "vrm_addon_extension", None), "spring_bone1", None
+    )
+    if not isinstance(spring_bone1, SpringBone1SpringBonePropertyGroup):
+        raise TypeError
     return spring_bone1
