@@ -5,11 +5,13 @@ import bpy
 from bpy.types import Armature, Bone, Object
 
 from ..common.logger import get_logger
-from .extension import VrmAddonArmatureExtensionPropertyGroup, get_armature_extension
-from .migration import migrate_all_objects
-from .mtoon1 import ops as mtoon1_ops
-from .vrm0.property_group import Vrm0HumanoidPropertyGroup
-from .vrm1 import property_group as vrm1_property_group
+from .migration import (
+    on_change_bpy_armature_name,
+    on_change_bpy_bone_name,
+    on_change_bpy_object_location,
+    on_change_bpy_object_mode,
+    on_change_bpy_object_name,
+)
 
 logger = get_logger(__name__)
 
@@ -92,56 +94,3 @@ def teardown_subscription() -> None:
     bpy.msgbus.clear_by_owner(subscription.object_location_subscription_owner)
     bpy.msgbus.clear_by_owner(subscription.object_mode_subscription_owner)
     bpy.msgbus.clear_by_owner(subscription.object_name_subscription_owner)
-
-
-def on_change_bpy_object_name() -> None:
-    context = bpy.context
-
-    for armature in context.blend_data.armatures:
-        ext = get_armature_extension(armature)
-        if (
-            tuple(ext.addon_version)
-            == VrmAddonArmatureExtensionPropertyGroup.INITIAL_ADDON_VERSION
-        ):
-            continue
-
-        # TODO: Needs optimization!
-        for collider in ext.spring_bone1.colliders:
-            collider.broadcast_bpy_object_name()
-
-
-def on_change_bpy_object_mode() -> None:
-    context = bpy.context
-
-    active_object = context.active_object
-    if not active_object:
-        return
-    mtoon1_ops.VRM_OT_refresh_mtoon1_outline.refresh_object(context, active_object)
-
-
-def on_change_bpy_object_location() -> None:
-    context = bpy.context
-    active_object = context.active_object
-    if not active_object:
-        return
-    vrm1_property_group.Vrm1LookAtPropertyGroup.update_all_previews(context)
-
-
-def on_change_bpy_bone_name() -> None:
-    context = bpy.context
-
-    for armature in context.blend_data.armatures:
-        ext = get_armature_extension(armature)
-        if (
-            tuple(ext.addon_version)
-            == VrmAddonArmatureExtensionPropertyGroup.INITIAL_ADDON_VERSION
-        ):
-            continue
-
-        Vrm0HumanoidPropertyGroup.update_all_node_candidates(context, armature.name)
-
-
-def on_change_bpy_armature_name() -> None:
-    context = bpy.context
-
-    migrate_all_objects(context, skip_non_migrated_armatures=True)
