@@ -1,26 +1,14 @@
 # SPDX-License-Identifier: MIT OR GPL-3.0-or-later
 import uuid
-import warnings
 from collections.abc import Set as AbstractSet
 from typing import TYPE_CHECKING
 
-from bpy.app.translations import pgettext
 from bpy.props import IntProperty, StringProperty
-from bpy.types import Armature, Context, Event, Object, Operator, UILayout
+from bpy.types import Armature, Context, Operator
 
 from ...common.human_bone_mapper.human_bone_mapper import create_human_bone_mapping
-from ...common.logger import get_logger
-from ...common.vrm0.human_bone import (
-    HumanBoneName,
-    HumanBoneSpecification,
-    HumanBoneSpecifications,
-)
 from ..extension import get_armature_extension
-from ..ops import VRM_OT_open_url_in_web_browser, layout_operator
-from ..property_group import BonePropertyGroup
 from .property_group import Vrm0HumanoidPropertyGroup
-
-logger = get_logger(__name__)
 
 
 class VRM_OT_add_vrm0_first_person_mesh_annotation(Operator):
@@ -29,31 +17,12 @@ class VRM_OT_add_vrm0_first_person_mesh_annotation(Operator):
     bl_description = "Add VRM 0.x First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -69,7 +38,6 @@ class VRM_OT_add_vrm0_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
 
 
@@ -79,36 +47,16 @@ class VRM_OT_remove_vrm0_first_person_mesh_annotation(Operator):
     bl_description = "Remove VRM 0.x First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     mesh_annotation_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -127,7 +75,6 @@ class VRM_OT_remove_vrm0_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         mesh_annotation_index: int  # type: ignore[no-redef]
 
@@ -138,36 +85,16 @@ class VRM_OT_move_up_vrm0_first_person_mesh_annotation(Operator):
     bl_description = "Move Up VRM 0.x First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     mesh_annotation_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -188,7 +115,6 @@ class VRM_OT_move_up_vrm0_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         mesh_annotation_index: int  # type: ignore[no-redef]
 
@@ -199,36 +125,16 @@ class VRM_OT_move_down_vrm0_first_person_mesh_annotation(Operator):
     bl_description = "Move Down VRM 0.x First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     mesh_annotation_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -247,7 +153,6 @@ class VRM_OT_move_down_vrm0_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         mesh_annotation_index: int  # type: ignore[no-redef]
 
@@ -258,36 +163,16 @@ class VRM_OT_add_vrm0_material_value_bind(Operator):
     bl_description = "Add VRM 0.x Blend Shape Material Value Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -308,7 +193,6 @@ class VRM_OT_add_vrm0_material_value_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
 
@@ -319,27 +203,9 @@ class VRM_OT_remove_vrm0_material_value_bind(Operator):
     bl_description = "Remove VRM 0.x Blend Shape Material Value Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -350,9 +216,7 @@ class VRM_OT_remove_vrm0_material_value_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -376,7 +240,6 @@ class VRM_OT_remove_vrm0_material_value_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         material_value_index: int  # type: ignore[no-redef]
@@ -388,27 +251,9 @@ class VRM_OT_move_up_vrm0_material_value_bind(Operator):
     bl_description = "Move Up VRM 0.x Blend Shape Material Value Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -419,9 +264,7 @@ class VRM_OT_move_up_vrm0_material_value_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -445,7 +288,6 @@ class VRM_OT_move_up_vrm0_material_value_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         material_value_index: int  # type: ignore[no-redef]
@@ -457,27 +299,9 @@ class VRM_OT_move_down_vrm0_material_value_bind(Operator):
     bl_description = "Move Down VRM 0.x Blend Shape Material Value Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -488,9 +312,7 @@ class VRM_OT_move_down_vrm0_material_value_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -514,7 +336,6 @@ class VRM_OT_move_down_vrm0_material_value_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         material_value_index: int  # type: ignore[no-redef]
@@ -526,27 +347,9 @@ class VRM_OT_add_vrm0_material_value_bind_target_value(Operator):
     bl_description = "Add VRM 0.x Blend Shape Material Value Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -557,9 +360,7 @@ class VRM_OT_add_vrm0_material_value_bind_target_value(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -581,7 +382,6 @@ class VRM_OT_add_vrm0_material_value_bind_target_value(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         material_value_index: int  # type: ignore[no-redef]
@@ -593,27 +393,9 @@ class VRM_OT_remove_vrm0_material_value_bind_target_value(Operator):
     bl_description = "Remove VRM 0.x Blend Shape Material Value Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -628,9 +410,7 @@ class VRM_OT_remove_vrm0_material_value_bind_target_value(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -655,7 +435,6 @@ class VRM_OT_remove_vrm0_material_value_bind_target_value(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         material_value_index: int  # type: ignore[no-redef]
@@ -668,36 +447,16 @@ class VRM_OT_add_vrm0_blend_shape_bind(Operator):
     bl_description = "Add VRM 0.x Blend Shape Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -716,7 +475,6 @@ class VRM_OT_add_vrm0_blend_shape_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
 
@@ -727,27 +485,9 @@ class VRM_OT_remove_vrm0_blend_shape_bind(Operator):
     bl_description = "Remove VRM 0.x Blend Shape Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -758,9 +498,7 @@ class VRM_OT_remove_vrm0_blend_shape_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -784,7 +522,6 @@ class VRM_OT_remove_vrm0_blend_shape_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -796,27 +533,9 @@ class VRM_OT_move_up_vrm0_blend_shape_bind(Operator):
     bl_description = "Move Up VRM 0.x Blend Shape Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -827,9 +546,7 @@ class VRM_OT_move_up_vrm0_blend_shape_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -851,7 +568,6 @@ class VRM_OT_move_up_vrm0_blend_shape_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -863,27 +579,9 @@ class VRM_OT_move_down_vrm0_blend_shape_bind(Operator):
     bl_description = "Move Up VRM 0.x Blend Shape Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -894,9 +592,7 @@ class VRM_OT_move_down_vrm0_blend_shape_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -918,39 +614,9 @@ class VRM_OT_move_down_vrm0_blend_shape_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
-
-
-class VRM_OT_restore_vrm0_blend_shape_group_bind_object(Operator):
-    bl_idname = "vrm.restore_vrm0_blend_shape_group_bind_object"
-    bl_label = "Restore Shape Key Assignments"
-    bl_description = "Restore VRM 0.x Blend Shape Group Bind Object Assignments"
-    bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
-
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def execute(self, context: Context) -> set[str]:
-        armature = context.blend_data.objects.get(self.armature_object_name)
-        if armature is None or armature.type != "ARMATURE":
-            return {"CANCELLED"}
-        armature_data = armature.data
-        if not isinstance(armature_data, Armature):
-            return {"CANCELLED"}
-        blend_shape_master = get_armature_extension(
-            armature_data
-        ).vrm0.blend_shape_master
-        blend_shape_master.restore_blend_shape_group_bind_object_assignments(context)
-        return {"FINISHED"}
-
-    if TYPE_CHECKING:
-        # This code is auto generated.
-        # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
 
 
 class VRM_OT_add_vrm0_secondary_animation_collider_group_collider(Operator):
@@ -959,27 +625,9 @@ class VRM_OT_add_vrm0_secondary_animation_collider_group_collider(Operator):
     bl_description = "Add VRM 0.x Collider"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     collider_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -989,9 +637,7 @@ class VRM_OT_add_vrm0_secondary_animation_collider_group_collider(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1011,7 +657,7 @@ class VRM_OT_add_vrm0_secondary_animation_collider_group_collider(Operator):
         if bone_name:
             collider_name = f"{bone_name}_collider"
         else:
-            collider_name = f"{self.armature_object_name}_collider"
+            collider_name = f"{self.armature_name}_collider"
 
         obj = context.blend_data.objects.new(name=collider_name, object_data=None)
         collider.bpy_object = obj
@@ -1030,7 +676,6 @@ class VRM_OT_add_vrm0_secondary_animation_collider_group_collider(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
         bone_name: str  # type: ignore[no-redef]
@@ -1042,27 +687,9 @@ class VRM_OT_remove_vrm0_secondary_animation_collider_group_collider(Operator):
     bl_description = "Remove VRM 0.x Collider"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     collider_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1073,9 +700,7 @@ class VRM_OT_remove_vrm0_secondary_animation_collider_group_collider(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1103,7 +728,6 @@ class VRM_OT_remove_vrm0_secondary_animation_collider_group_collider(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
         collider_index: int  # type: ignore[no-redef]
@@ -1115,27 +739,9 @@ class VRM_OT_move_up_vrm0_secondary_animation_collider_group_collider(Operator):
     bl_description = "Move Up VRM 0.x Collider"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     collider_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1146,9 +752,7 @@ class VRM_OT_move_up_vrm0_secondary_animation_collider_group_collider(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1170,7 +774,6 @@ class VRM_OT_move_up_vrm0_secondary_animation_collider_group_collider(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
         collider_index: int  # type: ignore[no-redef]
@@ -1182,27 +785,9 @@ class VRM_OT_move_down_vrm0_secondary_animation_collider_group_collider(Operator
     bl_description = "Move Down VRM 0.x Collider"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     collider_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1213,9 +798,7 @@ class VRM_OT_move_down_vrm0_secondary_animation_collider_group_collider(Operator
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1237,7 +820,6 @@ class VRM_OT_move_down_vrm0_secondary_animation_collider_group_collider(Operator
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
         collider_index: int  # type: ignore[no-redef]
@@ -1249,36 +831,16 @@ class VRM_OT_add_vrm0_secondary_animation_group_bone(Operator):
     bl_description = "Add VRM 0.x Secondary Animation Group Bone"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1290,14 +852,14 @@ class VRM_OT_add_vrm0_secondary_animation_group_bone(Operator):
         if len(bone_groups) <= self.bone_group_index:
             return {"CANCELLED"}
         bone_group = bone_groups[self.bone_group_index]
-        bone_group.bones.add()
+        bone = bone_group.bones.add()
+        bone.armature_data_name = armature_data.name
         bone_group.active_bone_index = len(bone_group.bones) - 1
         return {"FINISHED"}
 
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
 
@@ -1308,27 +870,9 @@ class VRM_OT_remove_vrm0_secondary_animation_group_bone(Operator):
     bl_description = "Remove VRM 0.x Secondary Animation Group Bone"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1339,9 +883,7 @@ class VRM_OT_remove_vrm0_secondary_animation_group_bone(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1364,7 +906,6 @@ class VRM_OT_remove_vrm0_secondary_animation_group_bone(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
         bone_index: int  # type: ignore[no-redef]
@@ -1376,27 +917,9 @@ class VRM_OT_move_up_vrm0_secondary_animation_group_bone(Operator):
     bl_description = "Move Up VRM 0.x Secondary Animation Group Bone"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1407,9 +930,7 @@ class VRM_OT_move_up_vrm0_secondary_animation_group_bone(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1431,7 +952,6 @@ class VRM_OT_move_up_vrm0_secondary_animation_group_bone(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
         bone_index: int  # type: ignore[no-redef]
@@ -1443,27 +963,9 @@ class VRM_OT_move_down_vrm0_secondary_animation_group_bone(Operator):
     bl_description = "Move Down VRM 0.x Secondary Animation Group Bone"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1474,9 +976,7 @@ class VRM_OT_move_down_vrm0_secondary_animation_group_bone(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1498,7 +998,6 @@ class VRM_OT_move_down_vrm0_secondary_animation_group_bone(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
         bone_index: int  # type: ignore[no-redef]
@@ -1510,36 +1009,16 @@ class VRM_OT_add_vrm0_secondary_animation_group_collider_group(Operator):
     bl_description = "Add VRM 0.x Secondary Animation Group Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1559,7 +1038,6 @@ class VRM_OT_add_vrm0_secondary_animation_group_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
 
@@ -1570,27 +1048,9 @@ class VRM_OT_remove_vrm0_secondary_animation_group_collider_group(Operator):
     bl_description = "Remove VRM 0.x Secondary Animation Group Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1601,9 +1061,7 @@ class VRM_OT_remove_vrm0_secondary_animation_group_collider_group(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1627,7 +1085,6 @@ class VRM_OT_remove_vrm0_secondary_animation_group_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
@@ -1639,27 +1096,9 @@ class VRM_OT_move_up_vrm0_secondary_animation_group_collider_group(Operator):
     bl_description = "Move Up VRM 0.x Secondary Animation Group Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1670,9 +1109,7 @@ class VRM_OT_move_up_vrm0_secondary_animation_group_collider_group(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1694,7 +1131,6 @@ class VRM_OT_move_up_vrm0_secondary_animation_group_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
@@ -1706,27 +1142,9 @@ class VRM_OT_move_down_vrm0_secondary_animation_group_collider_group(Operator):
     bl_description = "Move Down VRM 0.x Secondary Animation Group Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
@@ -1737,9 +1155,7 @@ class VRM_OT_move_down_vrm0_secondary_animation_group_collider_group(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1761,7 +1177,6 @@ class VRM_OT_move_down_vrm0_secondary_animation_group_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
@@ -1773,35 +1188,15 @@ class VRM_OT_add_vrm0_blend_shape_group(Operator):
     bl_description = "Add VRM 0.x Blend Shape Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"}
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1820,7 +1215,6 @@ class VRM_OT_add_vrm0_blend_shape_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         name: str  # type: ignore[no-redef]
 
@@ -1831,36 +1225,16 @@ class VRM_OT_remove_vrm0_blend_shape_group(Operator):
     bl_description = "Remove VRM 0.x Blend Shape Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1881,7 +1255,6 @@ class VRM_OT_remove_vrm0_blend_shape_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
 
@@ -1892,36 +1265,16 @@ class VRM_OT_move_up_vrm0_blend_shape_group(Operator):
     bl_description = "Move Up VRM 0.x Blend Shape Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1944,7 +1297,6 @@ class VRM_OT_move_up_vrm0_blend_shape_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
 
@@ -1955,36 +1307,16 @@ class VRM_OT_move_down_vrm0_blend_shape_group(Operator):
     bl_description = "Move Down VRM 0.x Blend Shape Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2007,7 +1339,6 @@ class VRM_OT_move_down_vrm0_blend_shape_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
 
@@ -2018,26 +1349,9 @@ class VRM_OT_add_vrm0_secondary_animation_group(Operator):
     bl_description = "Add VRM 0.x Secondary Animation Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     # Unnecessary property. Please do not use.
     blend_shape_group_index: IntProperty(  # type: ignore[valid-type]
@@ -2046,9 +1360,7 @@ class VRM_OT_add_vrm0_secondary_animation_group(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2066,7 +1378,6 @@ class VRM_OT_add_vrm0_secondary_animation_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         blend_shape_group_index: int  # type: ignore[no-redef]
 
@@ -2077,36 +1388,16 @@ class VRM_OT_remove_vrm0_secondary_animation_group(Operator):
     bl_description = "Remove VRM 0.x Secondary Animation Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2127,7 +1418,6 @@ class VRM_OT_remove_vrm0_secondary_animation_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
 
@@ -2138,36 +1428,16 @@ class VRM_OT_move_up_vrm0_secondary_animation_group(Operator):
     bl_description = "Move Up VRM 0.x Secondary Animation Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2186,7 +1456,6 @@ class VRM_OT_move_up_vrm0_secondary_animation_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
 
@@ -2197,36 +1466,16 @@ class VRM_OT_move_down_vrm0_secondary_animation_group(Operator):
     bl_description = "Move Down VRM 0.x Secondary Animation Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     bone_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2245,7 +1494,6 @@ class VRM_OT_move_down_vrm0_secondary_animation_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         bone_group_index: int  # type: ignore[no-redef]
 
@@ -2256,31 +1504,12 @@ class VRM_OT_add_vrm0_secondary_animation_collider_group(Operator):
     bl_description = "Add VRM 0.x Secondary Animation Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2299,7 +1528,6 @@ class VRM_OT_add_vrm0_secondary_animation_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
 
 
@@ -2309,36 +1537,16 @@ class VRM_OT_remove_vrm0_secondary_animation_collider_group(Operator):
     bl_description = "Remove VRM 0.x Secondary Animation Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     collider_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2365,7 +1573,6 @@ class VRM_OT_remove_vrm0_secondary_animation_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
 
@@ -2376,36 +1583,16 @@ class VRM_OT_move_up_vrm0_secondary_animation_collider_group(Operator):
     bl_description = "Move Up VRM 0.x Secondary Animation Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     collider_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2426,7 +1613,6 @@ class VRM_OT_move_up_vrm0_secondary_animation_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
 
@@ -2437,36 +1623,16 @@ class VRM_OT_move_down_vrm0_secondary_animation_collider_group(Operator):
     bl_description = "Move Down VRM 0.x Secondary Animation Collider Group"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     collider_group_index: IntProperty(  # type: ignore[valid-type]
         min=0,
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2487,7 +1653,6 @@ class VRM_OT_move_down_vrm0_secondary_animation_collider_group(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         collider_group_index: int  # type: ignore[no-redef]
 
@@ -2498,31 +1663,12 @@ class VRM_OT_assign_vrm0_humanoid_human_bones_automatically(Operator):
     bl_description = "Assign VRM 0.x Humanoid Human Bones"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -2548,7 +1694,7 @@ class VRM_OT_assign_vrm0_humanoid_human_bones_automatically(Operator):
                     or bone_name not in human_bone.node_candidates
                 ):
                     continue
-                human_bone.node.bone_name = bone_name
+                human_bone.node.set_bone_name(bone_name)
                 break
 
         Vrm0HumanoidPropertyGroup.update_all_node_candidates(
@@ -2559,209 +1705,4 @@ class VRM_OT_assign_vrm0_humanoid_human_bones_automatically(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
-
-
-def draw_bone_prop_search(
-    layout: UILayout,
-    human_bone_specification: HumanBoneSpecification,
-    armature: Object,
-    *,
-    show_diagnostics: bool = True,
-) -> None:
-    armature_data = armature.data
-    if not isinstance(armature_data, Armature):
-        return
-
-    humanoid = get_armature_extension(armature_data).vrm0.humanoid
-
-    props = None
-    for human_bone in humanoid.human_bones:
-        if human_bone.bone == human_bone_specification.name.value:
-            props = human_bone
-            break
-    if not props:
-        return
-
-    row = layout.row(align=True)
-    row.prop_search(
-        props.node,
-        "bone_name",
-        props,
-        "node_candidates",
-        text="",
-        translate=False,
-        icon=human_bone_specification.icon,
-    )
-
-    if not show_diagnostics:
-        return
-
-    if props.node.bone_name and props.node.bone_name not in props.node_candidates:
-        icon = "ERROR"
-        row.alert = True
-    else:
-        icon = "VIEWZOOM"
-
-    op = layout_operator(
-        row,
-        VRM_OT_show_vrm0_bone_assignment_diagnostics,
-        text="",
-        translate=False,
-        icon=icon,
-    )
-    op.armature_object_name = armature.name
-    op.human_bone_name = human_bone_specification.name.value
-    row.separator(factor=0.5)
-
-
-class VRM_OT_show_vrm0_bone_assignment_diagnostics(Operator):
-    bl_idname = "vrm.show_vrm0_bone_assignment_diagnostics"
-    bl_label = "Bone Assignment Diagnostics"
-    bl_description = (
-        "Shows the cause of the bone assignment error"
-        " or the reason why none of the bone assignment candidates exist."
-    )
-
-    def execute(self, _context: Context) -> set[str]:
-        return {"FINISHED"}
-
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"}
-    )
-
-    human_bone_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"}
-    )
-
-    def invoke(self, context: Context, _event: Event) -> set[str]:
-        obj = context.blend_data.objects.get(self.armature_object_name)
-        if not obj or obj.type != "ARMATURE":
-            return {"CANCELLED"}
-
-        if not HumanBoneName.from_str(self.human_bone_name):
-            return {"CANCELLED"}
-
-        return context.window_manager.invoke_props_dialog(self, width=800)
-
-    def draw(self, context: Context) -> None:
-        armature = context.blend_data.objects.get(self.armature_object_name)
-        if not armature:
-            return
-        armature_data = armature.data
-        if not isinstance(armature_data, Armature):
-            return
-        human_bone_name = HumanBoneName.from_str(self.human_bone_name)
-        if not human_bone_name:
-            return
-        human_bone_specification = HumanBoneSpecifications.get(human_bone_name)
-
-        layout = self.layout
-        head_row = layout.row()
-        head_row.alignment = "LEFT"
-        head_row.label(
-            text=pgettext('Assignment of the VRM Human Bone "{vrm_human_bone}"').format(
-                vrm_human_bone=human_bone_specification.title
-            ),
-            icon="ARMATURE_DATA",
-        )
-        draw_bone_prop_search(
-            head_row, human_bone_specification, armature, show_diagnostics=False
-        )
-        help_op = layout_operator(
-            head_row,
-            VRM_OT_open_url_in_web_browser,
-            icon="URL",
-            text="Open help in a Web Browser",
-        )
-        help_op.url = pgettext(
-            "https://github.com/vrm-c/vrm-specification/blob/c24d76d99a18738dd2c266be1c83f089064a7b5e/specification/VRMC_vrm-1.0/humanoid.md#humanoid-bone-parent-child-relationship"
-        )
-
-        ext = get_armature_extension(armature_data)
-        humanoid = ext.vrm0.humanoid
-        human_bone = next(
-            (
-                human_bone
-                for human_bone in humanoid.human_bones
-                if human_bone.bone == human_bone_name.value
-            ),
-            None,
-        )
-        if not human_bone:
-            return
-        bone_name = human_bone.node.bone_name
-        if bone_name in human_bone.node_candidates:
-            layout.label(
-                text=pgettext(
-                    'The bone "{bone_name}" of the armature "{armature_object_name}"'
-                    + ' can be assigned to the VRM Human Bone "{human_bone_name}".'
-                ).format(
-                    armature_object_name=self.armature_object_name,
-                    bone_name=bone_name,
-                    human_bone_name=human_bone_specification.title,
-                ),
-                translate=False,
-                icon="CHECKMARK",
-            )
-        elif bone_name:
-            layout.label(
-                text=pgettext(
-                    'The bone "{bone_name}" of the armature "{armature_object_name}"'
-                    + ' cannot be assigned to the VRM Human Bone "{human_bone_name}".'
-                ).format(
-                    armature_object_name=self.armature_object_name,
-                    bone_name=bone_name,
-                    human_bone_name=human_bone_specification.title,
-                ),
-                translate=False,
-                icon="ERROR",
-            )
-        elif not human_bone.node_candidates:
-            layout.label(
-                text=pgettext(
-                    'The armature "{armature_object_name}" does not have any bones'
-                    + ' that can be assigned to the VRM Human Bone "{human_bone_name}".'
-                ).format(
-                    armature_object_name=self.armature_object_name,
-                    human_bone_name=human_bone_specification.title,
-                ),
-                translate=False,
-                icon="ERROR",
-            )
-
-        bpy_bone_name_to_human_bone_specification: dict[
-            str, HumanBoneSpecification
-        ] = {}
-        error_bpy_bone_names = list[str]()
-
-        for human_bone in humanoid.human_bones:
-            human_bone_name = HumanBoneName.from_str(human_bone.bone)
-            if not human_bone_name:
-                continue
-            bpy_bone_name = human_bone.node.bone_name
-            if not bpy_bone_name:
-                continue
-            if bpy_bone_name not in human_bone.node_candidates:
-                error_bpy_bone_names.append(bpy_bone_name)
-            bpy_bone_name_to_human_bone_specification[bpy_bone_name] = (
-                HumanBoneSpecifications.get(human_bone_name)
-            )
-
-        Vrm0HumanoidPropertyGroup.update_all_node_candidates(
-            context, armature_data.name
-        )
-        BonePropertyGroup.find_bone_candidates(
-            armature_data,
-            human_bone_specification,
-            bpy_bone_name_to_human_bone_specification,
-            error_bpy_bone_names,
-            layout,
-        )
-
-    if TYPE_CHECKING:
-        # This code is auto generated.
-        # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
-        human_bone_name: str  # type: ignore[no-redef]
