@@ -1,18 +1,15 @@
 # SPDX-License-Identifier: MIT OR GPL-3.0-or-later
 import re
-import warnings
 from collections.abc import Mapping, Sequence
 from collections.abc import Set as AbstractSet
 from typing import TYPE_CHECKING, ClassVar, Optional, Protocol
 
 import bpy
-from bpy.app.translations import pgettext
 from bpy.props import IntProperty, StringProperty
 from bpy.types import (
     ID,
     Armature,
     Context,
-    Event,
     FCurve,
     Material,
     Node,
@@ -26,21 +23,14 @@ from bpy.types import (
     ShaderNodeTree,
     ShaderNodeValue,
     ShaderNodeVectorMath,
-    UILayout,
 )
 
 from ...common import ops, shader
 from ...common.human_bone_mapper.human_bone_mapper import create_human_bone_mapping
 from ...common.logger import get_logger
 from ...common.vrm0.human_bone import HumanBoneName as Vrm0HumanBoneName
-from ...common.vrm1.human_bone import (
-    HumanBoneName,
-    HumanBoneSpecification,
-    HumanBoneSpecifications,
-)
+from ...common.vrm1.human_bone import HumanBoneName, HumanBoneSpecifications
 from ..extension import get_armature_extension
-from ..ops import VRM_OT_open_url_in_web_browser, layout_operator
-from ..property_group import BonePropertyGroup
 from ..vrm0.property_group import Vrm0HumanoidPropertyGroup
 from .property_group import (
     Vrm1ExpressionPropertyGroup,
@@ -56,31 +46,12 @@ class VRM_OT_add_vrm1_meta_author(Operator):
     bl_description = "Add VRM 1.0 Meta Author"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -95,7 +66,6 @@ class VRM_OT_add_vrm1_meta_author(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
 
 
@@ -105,36 +75,16 @@ class VRM_OT_remove_vrm1_meta_author(Operator):
     bl_description = "Remove VRM 1.0 Meta Author"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     author_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -153,7 +103,6 @@ class VRM_OT_remove_vrm1_meta_author(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         author_index: int  # type: ignore[no-redef]
 
@@ -164,36 +113,16 @@ class VRM_OT_move_up_vrm1_meta_author(Operator):
     bl_description = "Move Up VRM 1.0 Meta Author"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     author_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -210,7 +139,6 @@ class VRM_OT_move_up_vrm1_meta_author(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         author_index: int  # type: ignore[no-redef]
 
@@ -221,36 +149,16 @@ class VRM_OT_move_down_vrm1_meta_author(Operator):
     bl_description = "Move Down VRM 1.0 Meta Author"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     author_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -267,7 +175,6 @@ class VRM_OT_move_down_vrm1_meta_author(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         author_index: int  # type: ignore[no-redef]
 
@@ -278,31 +185,12 @@ class VRM_OT_add_vrm1_meta_reference(Operator):
     bl_description = "Add VRM 1.0 Meta Reference"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -317,7 +205,6 @@ class VRM_OT_add_vrm1_meta_reference(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
 
 
@@ -327,36 +214,16 @@ class VRM_OT_remove_vrm1_meta_reference(Operator):
     bl_description = "Remove VRM 1.0 Meta Reference"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     reference_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -375,7 +242,6 @@ class VRM_OT_remove_vrm1_meta_reference(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         reference_index: int  # type: ignore[no-redef]
 
@@ -386,36 +252,16 @@ class VRM_OT_move_up_vrm1_meta_reference(Operator):
     bl_description = "Move Up VRM 1.0 Meta Reference"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     reference_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -432,7 +278,6 @@ class VRM_OT_move_up_vrm1_meta_reference(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         reference_index: int  # type: ignore[no-redef]
 
@@ -443,36 +288,16 @@ class VRM_OT_move_down_vrm1_meta_reference(Operator):
     bl_description = "Move Down VRM 1.0 Meta Reference"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     reference_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -489,7 +314,6 @@ class VRM_OT_move_down_vrm1_meta_reference(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         reference_index: int  # type: ignore[no-redef]
 
@@ -500,35 +324,15 @@ class VRM_OT_add_vrm1_expressions_custom_expression(Operator):
     bl_description = "Add VRM 1.0 Custom Expression"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     custom_expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -546,7 +350,6 @@ class VRM_OT_add_vrm1_expressions_custom_expression(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         custom_expression_name: str  # type: ignore[no-redef]
 
@@ -557,35 +360,15 @@ class VRM_OT_remove_vrm1_expressions_custom_expression(Operator):
     bl_description = "Remove VRM 1.0 Custom Expression"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     custom_expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -607,7 +390,6 @@ class VRM_OT_remove_vrm1_expressions_custom_expression(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         custom_expression_name: str  # type: ignore[no-redef]
 
@@ -618,35 +400,15 @@ class VRM_OT_move_up_vrm1_expressions_custom_expression(Operator):
     bl_description = "Move Up VRM 1.0 Custom Expression"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     custom_expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -673,7 +435,6 @@ class VRM_OT_move_up_vrm1_expressions_custom_expression(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         custom_expression_name: str  # type: ignore[no-redef]
 
@@ -684,35 +445,15 @@ class VRM_OT_move_down_vrm1_expressions_custom_expression(Operator):
     bl_description = "Move Down VRM 1.0 Custom Expression"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     custom_expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -739,7 +480,6 @@ class VRM_OT_move_down_vrm1_expressions_custom_expression(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         custom_expression_name: str  # type: ignore[no-redef]
 
@@ -750,31 +490,12 @@ class VRM_OT_add_vrm1_first_person_mesh_annotation(Operator):
     bl_description = "Add VRM 1.0 First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -790,7 +511,6 @@ class VRM_OT_add_vrm1_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
 
 
@@ -800,36 +520,16 @@ class VRM_OT_remove_vrm1_first_person_mesh_annotation(Operator):
     bl_description = "Remove VRM 1.0 First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     mesh_annotation_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -848,7 +548,6 @@ class VRM_OT_remove_vrm1_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         mesh_annotation_index: int  # type: ignore[no-redef]
 
@@ -859,36 +558,16 @@ class VRM_OT_move_up_vrm1_first_person_mesh_annotation(Operator):
     bl_description = "Move Up VRM 1.0 First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     mesh_annotation_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -907,7 +586,6 @@ class VRM_OT_move_up_vrm1_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         mesh_annotation_index: int  # type: ignore[no-redef]
 
@@ -918,36 +596,16 @@ class VRM_OT_move_down_vrm1_first_person_mesh_annotation(Operator):
     bl_description = "Move Down VRM 1.0 First Person Mesh Annotation"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     mesh_annotation_index: IntProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
         min=0,
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -966,7 +624,6 @@ class VRM_OT_move_down_vrm1_first_person_mesh_annotation(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         mesh_annotation_index: int  # type: ignore[no-redef]
 
@@ -977,35 +634,15 @@ class VRM_OT_add_vrm1_expression_morph_target_bind(Operator):
     bl_description = "Add VRM 1.0 Expression Morph Target Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1024,7 +661,6 @@ class VRM_OT_add_vrm1_expression_morph_target_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
 
@@ -1035,27 +671,9 @@ class VRM_OT_remove_vrm1_expression_morph_target_bind(Operator):
     bl_description = "Remove VRM 1.0 Expression Morph Target Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1065,9 +683,7 @@ class VRM_OT_remove_vrm1_expression_morph_target_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1089,7 +705,6 @@ class VRM_OT_remove_vrm1_expression_morph_target_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1101,27 +716,9 @@ class VRM_OT_move_up_vrm1_expression_morph_target_bind(Operator):
     bl_description = "Move Up VRM 1.0 Expression Morph Target Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1131,9 +728,7 @@ class VRM_OT_move_up_vrm1_expression_morph_target_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1153,7 +748,6 @@ class VRM_OT_move_up_vrm1_expression_morph_target_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1165,27 +759,9 @@ class VRM_OT_move_down_vrm1_expression_morph_target_bind(Operator):
     bl_description = "Move Down VRM 1.0 Expression Morph Target Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1195,9 +771,7 @@ class VRM_OT_move_down_vrm1_expression_morph_target_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1217,37 +791,9 @@ class VRM_OT_move_down_vrm1_expression_morph_target_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
-
-
-class VRM_OT_restore_vrm1_expression_morph_target_bind_object(Operator):
-    bl_idname = "vrm.restore_vrm1_expression_morph_target_bind_object"
-    bl_label = "Restore Shape Key Assignments"
-    bl_description = "Restore VRM 1.0 Expression Morph Target Bind Object Assignments"
-    bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
-
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def execute(self, context: Context) -> set[str]:
-        armature = context.blend_data.objects.get(self.armature_object_name)
-        if armature is None or armature.type != "ARMATURE":
-            return {"CANCELLED"}
-        armature_data = armature.data
-        if not isinstance(armature_data, Armature):
-            return {"CANCELLED"}
-        expressions = get_armature_extension(armature_data).vrm1.expressions
-        expressions.restore_expression_morph_target_bind_object_assignments(context)
-        return {"FINISHED"}
-
-    if TYPE_CHECKING:
-        # This code is auto generated.
-        # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
 
 
 class VRM_OT_add_vrm1_expression_material_color_bind(Operator):
@@ -1256,35 +802,15 @@ class VRM_OT_add_vrm1_expression_material_color_bind(Operator):
     bl_description = "Add VRM 1.0 Expression Material Value Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1305,7 +831,6 @@ class VRM_OT_add_vrm1_expression_material_color_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
 
@@ -1316,27 +841,9 @@ class VRM_OT_remove_vrm1_expression_material_color_bind(Operator):
     bl_description = "Remove VRM 1.0 Expression Material Color Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1346,9 +853,7 @@ class VRM_OT_remove_vrm1_expression_material_color_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1370,7 +875,6 @@ class VRM_OT_remove_vrm1_expression_material_color_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1382,27 +886,9 @@ class VRM_OT_move_up_vrm1_expression_material_color_bind(Operator):
     bl_description = "Move Up VRM 1.0 Expression Material Color Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1412,9 +898,7 @@ class VRM_OT_move_up_vrm1_expression_material_color_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1434,7 +918,6 @@ class VRM_OT_move_up_vrm1_expression_material_color_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1446,27 +929,9 @@ class VRM_OT_move_down_vrm1_expression_material_color_bind(Operator):
     bl_description = "Move Down VRM 1.0 Expression Material Color Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1476,9 +941,7 @@ class VRM_OT_move_down_vrm1_expression_material_color_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1498,7 +961,6 @@ class VRM_OT_move_down_vrm1_expression_material_color_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1510,35 +972,15 @@ class VRM_OT_add_vrm1_expression_texture_transform_bind(Operator):
     bl_description = "Add VRM 1.0 Expression Texture Transform Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1557,7 +999,6 @@ class VRM_OT_add_vrm1_expression_texture_transform_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
 
@@ -1568,27 +1009,9 @@ class VRM_OT_remove_vrm1_expression_texture_transform_bind(Operator):
     bl_description = "Remove VRM 1.0 Expression Texture Transform Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1598,9 +1021,7 @@ class VRM_OT_remove_vrm1_expression_texture_transform_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1622,7 +1043,6 @@ class VRM_OT_remove_vrm1_expression_texture_transform_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1634,27 +1054,9 @@ class VRM_OT_move_up_vrm1_expression_texture_transform_bind(Operator):
     bl_description = "Move Up VRM 1.0 Expression Texture Transform Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1664,9 +1066,7 @@ class VRM_OT_move_up_vrm1_expression_texture_transform_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1686,7 +1086,6 @@ class VRM_OT_move_up_vrm1_expression_texture_transform_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1698,27 +1097,9 @@ class VRM_OT_move_down_vrm1_expression_texture_transform_bind(Operator):
     bl_description = "Move Down VRM 1.0 Expression Morph Target Bind"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
     )
@@ -1728,9 +1109,7 @@ class VRM_OT_move_down_vrm1_expression_texture_transform_bind(Operator):
     )
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1750,7 +1129,6 @@ class VRM_OT_move_down_vrm1_expression_texture_transform_bind(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
         bind_index: int  # type: ignore[no-redef]
@@ -1768,31 +1146,12 @@ class VRM_OT_assign_vrm1_humanoid_human_bones_automatically(Operator):
     bl_description = "Assign VRM 1.0 Humanoid Human Bones"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"},
-        update=update_armature_name,
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1831,7 +1190,7 @@ class VRM_OT_assign_vrm1_humanoid_human_bones_automatically(Operator):
                     continue
                 if vrm0_human_bone.node.bone_name not in human_bone.node_candidates:
                     continue
-                human_bone.node.bone_name = vrm0_human_bone.node.bone_name
+                human_bone.node.set_bone_name(vrm0_human_bone.node.bone_name)
 
         for (
             bone_name,
@@ -1848,7 +1207,7 @@ class VRM_OT_assign_vrm1_humanoid_human_bones_automatically(Operator):
                     or bone_name not in human_bone.node_candidates
                 ):
                     continue
-                human_bone.node.bone_name = bone_name
+                human_bone.node.set_bone_name(bone_name)
                 break
 
         Vrm1HumanBonesPropertyGroup.update_all_node_candidates(
@@ -1859,7 +1218,6 @@ class VRM_OT_assign_vrm1_humanoid_human_bones_automatically(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
 
 
@@ -1899,27 +1257,9 @@ class VRM_OT_refresh_vrm1_expression_texture_transform_bind_preview(Operator):
     bl_description = "Refresh VRM 1.0 Expression Texture Transform Bind Preview"
     bl_options: AbstractSet[str] = {"REGISTER", "UNDO"}
 
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-    )
-
-    def update_armature_name(self, _context: Context) -> None:
-        message = (
-            f"`{type(self).__qualname__}.armature_name` is deprecated"
-            + " and will be removed in the next major release."
-            + f" `Please use {type(self).__qualname__}.armature_object_name` instead."
-        )
-        logger.warning(message)
-        warnings.warn(message, DeprecationWarning, stacklevel=5)
-
     armature_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"},
-        update=update_armature_name,
+        options={"HIDDEN"}
     )
-    """`armature_name` is deprecated and will be removed in the next major release.
-    Please use `armature_object_name` instead.
-    """
-
     expression_name: StringProperty(  # type: ignore[valid-type]
         options={"HIDDEN"}
     )
@@ -1946,9 +1286,7 @@ class VRM_OT_refresh_vrm1_expression_texture_transform_bind_preview(Operator):
     }
 
     def execute(self, context: Context) -> set[str]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = context.blend_data.objects.get(self.armature_object_name)
+        armature = context.blend_data.objects.get(self.armature_name)
         if armature is None or armature.type != "ARMATURE":
             return {"CANCELLED"}
         armature_data = armature.data
@@ -1974,24 +1312,28 @@ class VRM_OT_refresh_vrm1_expression_texture_transform_bind_preview(Operator):
         return {"FINISHED"}
 
     def get_all_expressions(self) -> list[tuple[Vrm1ExpressionPropertyGroup, str, str]]:
-        if not self.armature_object_name and self.armature_name:
-            self.armature_object_name = self.armature_name
-        armature = bpy.data.objects.get(self.armature_object_name)
+        armature = bpy.data.objects.get(self.armature_name)
         if not armature or armature.type != "ARMATURE":
             return []
-        armature_data = armature.data
-        if not isinstance(armature_data, Armature):
+
+        extension = getattr(armature.data, "vrm_addon_extension", None)
+        if (
+            extension is not None
+            and extension.vrm1 is not None
+            and extension.vrm1.expressions is not None
+        ):
+            expressions = extension.vrm1.expressions
+        else:
             return []
-        ext = get_armature_extension(armature_data)
-        expressions = ext.vrm1.expressions
 
         all_expressions: list[tuple[Vrm1ExpressionPropertyGroup, str, str]] = []
-        for preset_name in self.preset_name_mapping.values():
-            preset_expr = getattr(expressions.preset, preset_name, None)
-            if preset_expr:
-                all_expressions.append((preset_expr, "preset", preset_name))
-        for i, custom_expr in enumerate(expressions.custom):
-            all_expressions.append((custom_expr, "custom", str(i)))
+        if expressions is not None:
+            for preset_name in self.preset_name_mapping.values():
+                preset_expr = getattr(expressions.preset, preset_name, None)
+                if preset_expr:
+                    all_expressions.append((preset_expr, "preset", preset_name))
+            for i, custom_expr in enumerate(expressions.custom):
+                all_expressions.append((custom_expr, "custom", str(i)))
 
         return all_expressions
 
@@ -2479,8 +1821,8 @@ class VRM_OT_refresh_vrm1_expression_texture_transform_bind_preview(Operator):
             elif (
                 node.type == "GROUP"
                 and isinstance(node, ShaderNodeGroup)
-                and (node_node_tree := node.node_tree)
-                and node_node_tree.name.startswith("VRM_TextureTransform_")
+                and node.node_tree
+                and node.node_tree.name.startswith("VRM_TextureTransform_")
             ):
                 nodes_to_remove.append(node)
                 # Handle links for group nodes
@@ -2501,7 +1843,7 @@ class VRM_OT_refresh_vrm1_expression_texture_transform_bind_preview(Operator):
 
         # Remove orphaned node groups
         for group in bpy.data.node_groups:
-            if group.name.startswith("VRM_TextureTransform_") and not group.users:
+            if group.name.startswith("VRM_TextureTransform_") and group.users == 0:
                 bpy.data.node_groups.remove(group)
 
     def connect_group_to_image_node(
@@ -2532,199 +1874,5 @@ class VRM_OT_refresh_vrm1_expression_texture_transform_bind_preview(Operator):
     if TYPE_CHECKING:
         # This code is auto generated.
         # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
         armature_name: str  # type: ignore[no-redef]
         expression_name: str  # type: ignore[no-redef]
-
-
-def draw_bone_prop_search(
-    layout: UILayout,
-    human_bone_specification: HumanBoneSpecification,
-    armature: Object,
-    *,
-    show_diagnostics: bool = True,
-) -> None:
-    armature_data = armature.data
-    if not isinstance(armature_data, Armature):
-        return
-    ext = get_armature_extension(armature_data)
-    human_bone = ext.vrm1.humanoid.human_bones.human_bone_name_to_human_bone().get(
-        human_bone_specification.name
-    )
-    if not human_bone:
-        return
-
-    row = layout.row(align=True)
-    row.prop_search(
-        human_bone.node,
-        "bone_name",
-        human_bone,
-        "node_candidates",
-        text="",
-        translate=False,
-        icon=human_bone_specification.icon,
-    )
-    if not show_diagnostics:
-        return
-
-    if (
-        human_bone.node.bone_name
-        and human_bone.node.bone_name not in human_bone.node_candidates
-    ):
-        icon = "ERROR"
-        row.alert = True
-    else:
-        icon = "VIEWZOOM"
-
-    op = layout_operator(
-        row,
-        VRM_OT_show_vrm1_bone_assignment_diagnostics,
-        text="",
-        translate=False,
-        icon=icon,
-    )
-    op.armature_object_name = armature.name
-    op.human_bone_name = human_bone_specification.name.value
-    row.separator(factor=0.5)
-
-
-class VRM_OT_show_vrm1_bone_assignment_diagnostics(Operator):
-    bl_idname = "vrm.show_vrm1_bone_assignment_diagnostics"
-    bl_label = "Bone Assignment Diagnostics"
-    bl_description = (
-        "Shows the cause of the bone assignment error"
-        " or the reason why none of the bone assignment candidates exist."
-    )
-
-    def execute(self, _context: Context) -> set[str]:
-        return {"FINISHED"}
-
-    armature_object_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"}
-    )
-
-    human_bone_name: StringProperty(  # type: ignore[valid-type]
-        options={"HIDDEN"}
-    )
-
-    def invoke(self, context: Context, _event: Event) -> set[str]:
-        obj = context.blend_data.objects.get(self.armature_object_name)
-        if not obj or obj.type != "ARMATURE":
-            return {"CANCELLED"}
-
-        if not HumanBoneName.from_str(self.human_bone_name):
-            return {"CANCELLED"}
-
-        return context.window_manager.invoke_props_dialog(self, width=800)
-
-    def draw(self, context: Context) -> None:
-        armature = context.blend_data.objects.get(self.armature_object_name)
-        if not armature:
-            return
-        armature_data = armature.data
-        if not isinstance(armature_data, Armature):
-            return
-        human_bone_name = HumanBoneName.from_str(self.human_bone_name)
-        if not human_bone_name:
-            return
-        human_bone_specification = HumanBoneSpecifications.get(human_bone_name)
-
-        layout = self.layout
-        head_row = layout.row()
-        head_row.alignment = "LEFT"
-        head_row.label(
-            text=pgettext('Assignment of the VRM Human Bone "{vrm_human_bone}"').format(
-                vrm_human_bone=human_bone_specification.title
-            ),
-            icon="ARMATURE_DATA",
-        )
-        draw_bone_prop_search(
-            head_row, human_bone_specification, armature, show_diagnostics=False
-        )
-        help_op = layout_operator(
-            head_row,
-            VRM_OT_open_url_in_web_browser,
-            icon="URL",
-            text="Open help in a Web Browser",
-        )
-        help_op.url = pgettext(
-            "https://github.com/vrm-c/vrm-specification/blob/c24d76d99a18738dd2c266be1c83f089064a7b5e/specification/VRMC_vrm-1.0/humanoid.md#humanoid-bone-parent-child-relationship"
-        )
-
-        ext = get_armature_extension(armature_data)
-        humanoid = ext.vrm1.humanoid
-        human_bone_name_to_human_bone = (
-            humanoid.human_bones.human_bone_name_to_human_bone()
-        )
-        human_bone = human_bone_name_to_human_bone.get(human_bone_name)
-        if not human_bone:
-            return
-
-        bone_name = human_bone.node.bone_name
-        if bone_name in human_bone.node_candidates:
-            layout.label(
-                text=pgettext(
-                    'The bone "{bone_name}" of the armature "{armature_object_name}"'
-                    + ' can be assigned to the VRM Human Bone "{human_bone_name}".'
-                ).format(
-                    armature_object_name=self.armature_object_name,
-                    bone_name=bone_name,
-                    human_bone_name=human_bone_specification.title,
-                ),
-                translate=False,
-                icon="CHECKMARK",
-            )
-        elif bone_name:
-            layout.label(
-                text=pgettext(
-                    'The bone "{bone_name}" of the armature "{armature_object_name}"'
-                    + ' cannot be assigned to the VRM Human Bone "{human_bone_name}".'
-                ).format(
-                    armature_object_name=self.armature_object_name,
-                    bone_name=bone_name,
-                    human_bone_name=human_bone_specification.title,
-                ),
-                translate=False,
-                icon="ERROR",
-            )
-        elif not human_bone.node_candidates:
-            layout.label(
-                text=pgettext(
-                    'The armature "{armature_object_name}" does not have any bones'
-                    + ' that can be assigned to the VRM Human Bone "{human_bone_name}".'
-                ).format(
-                    armature_object_name=self.armature_object_name,
-                    human_bone_name=human_bone_specification.title,
-                ),
-                translate=False,
-                icon="ERROR",
-            )
-
-        bpy_bone_name_to_human_bone_specification = {
-            bpy_bone_name: HumanBoneSpecifications.get(human_bone_name)
-            for human_bone_name, human_bone in human_bone_name_to_human_bone.items()
-            if (bpy_bone_name := human_bone.node.bone_name)
-        }
-        error_bpy_bone_names = [
-            bpy_bone_name
-            for human_bone in human_bone_name_to_human_bone.values()
-            if (bpy_bone_name := human_bone.node.bone_name)
-            and bpy_bone_name not in human_bone.node_candidates
-        ]
-
-        Vrm1HumanBonesPropertyGroup.update_all_node_candidates(
-            context, armature_data.name
-        )
-        BonePropertyGroup.find_bone_candidates(
-            armature_data,
-            human_bone_specification,
-            bpy_bone_name_to_human_bone_specification,
-            error_bpy_bone_names,
-            layout,
-        )
-
-    if TYPE_CHECKING:
-        # This code is auto generated.
-        # To regenerate, run the `uv run tools/property_typing.py` command.
-        armature_object_name: str  # type: ignore[no-redef]
-        human_bone_name: str  # type: ignore[no-redef]
